% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulateIncrArea.R
\name{accumulateIncrArea}
\alias{accumulateIncrArea}
\title{accumulateIncrArea}
\usage{
accumulateIncrArea(indata, accum_elements, accum_names)
}
\arguments{
\item{indata}{data.frame containing reach navigation variables and variables to be selected
for incremental area accumulation}

\item{accum_elements}{character vector indicating variables to be selected for incremental
area accumulation}

\item{accum_names}{character vector of final names for accumlated data}
}
\value{
\code{AreaRch}  data.frame containing waterid and \code{accum_elements}
}
\description{
Accumulates incremental area for \code{accum_elements} variables, expressed in areal
units. Function can only be run after replaceData1Names.R has been executed. Uses reach
navigation data. \cr \cr
Executed By: \itemize{\item calcClassLandusePercent.R
\item createVerifyReachAttr.R
\item hydseq.R} \cr
}
