% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMarkerText.R
\name{addMarkerText}
\alias{addMarkerText}
\title{addMarkerText}
\usage{
addMarkerText(markerText, add_plotlyVars, mapData, sourceData)
}
\arguments{
\item{markerText}{character string starting plotly markerText (example:
\code{markerText<-"~paste('</br> ',master_map_list[k],' :', round(mapdataname,predictionClassRounding)"})}

\item{add_plotlyVars}{character vector indicating user selected variables to add to plot hover
text}

\item{mapData}{data.frame of data to be plotted}

\item{sourceData}{data.frame with additional variables to be added to plot data for hover text}
}
\value{
\code{marker.list} a named list containing \code{markerText} the character string to execute to
generate plotly hover text and \code{mapData} the data to be plotted
}
\description{
Adds user selected \code{add_plotlyVars} to data to be mapped and creates the string
to execute in plotly to add to the hover text. \cr \cr
Executed By: \itemize{\item checkDrainageareaErrors.Rmd
\item checkDrainageareaErrorsChild.Rmd
\item diagnosticContiguousChild.Rmd
\item diagnosticCorrChild.Rmd
\item diagnosticMaps.R
\item diagnosticPlotsNLLS.Rmd
\item mapSiteAttributes.R
\item predictMaps.R
\item predictMaps.Rmd} \cr
Executes Routines: named.list.R \cr
}
