% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggDynamicMapdata.R
\name{aggDynamicMapdata}
\alias{aggDynamicMapdata}
\title{aggDynamicMapdata}
\usage{
aggDynamicMapdata(
  map_years,
  map_seasons,
  enable_plotlyMaps,
  add_plotlyVars,
  aggFuncs,
  vvar,
  MAPID,
  commonvar,
  subdata
)
}
\arguments{
\item{map_years}{RSPARROW dynamic user setting indicating how dynamic mapping
should proceed with regards to the year timestep.  A numeric
vector of years indicates mapping of each year in the vector.
A character string equal to "all" indicates that all years
present in the data1 file should be mapped. A character string
from the following list of approved aggregation functions
c("mean","median","min","max") indicates that the selected
aggregation function should be applied to the "year" attribute
to obtain a seasonal mean, median, min or max.  NA indicates that
the year timestep is not present in data1}

\item{map_seasons}{RSPARROW dynamic user setting indicating how dynamic mapping
should proceed with regards to the season timestep.  A character
vector of seasons indicates mapping of each season in the vector.
Seasons must be from the following list c("winter","spring","summer"
"fall").A character string equal to "all" indicates that all seasons
present in the data1 file should be mapped. A character string
from the following list of approved aggregation functions
c("mean","median","min","max") indicates that the selected
aggregation function should be applied to the "season" attribute
to obtain an annual mean, median, min or max.  NA indicates that
the season timestep is not present in data1}

\item{enable_plotlyMaps}{yes/no indicating whether to enable plotly interactive
displays for maps}

\item{add_plotlyVars}{character vector indicating user selected variables to add to plot hover
text}

\item{aggFuncs}{character vector of possible aggregate functions. For RSPARROW 2.0,
c("mean","median","min","max")}

\item{vvar}{mapping variable as vector}

\item{MAPID}{vector of waterid_for_RSPARROW_mapping}

\item{commonvar}{string indicating the column to join the map data and the shapefile}

\item{subdata}{data.frame input data (subdata)}
}
\value{
\code{map_agg.list} named list including data.frame aggregated according to \code{map_years}
and \code{map_seasons} and all objects required to generate a map.
}
\description{
Aggregate data for dynamic mapping by year, season, or year
and season based on the user selected \code{map_years} and \code{map_seasons}
settings.  Available aggregation functions include mean, median,
min, and max. \cr \cr
Executed By: \itemize{\item diagnosticPlotsNLLS.Rmd
\item diagnosticResidMapExec.R
\item goShinyPlot.R
\item predictMaps.R} \cr
Executes Routines: \itemize{\item test_addPlotlyvars.R
\item named.list.R} \cr
}
