% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allowRemoveRow.R
\name{allowRemoveRow}
\alias{allowRemoveRow}
\title{allowRemoveRow}
\usage{
allowRemoveRow(hot, allowRemove)
}
\arguments{
\item{hot}{shiny rhandsontable hot table used for user input}

\item{allowRemove}{TRUE/FALSE indicates whether the user should have the option to remove
rows from a rhandsontable hottable in shiny.  If only 1 row in the table, then FALSE}
}
\description{
Function to edit context menu for rhandsontable hottables in shiny ap, so that
the first row cannot be deleted. \cr \cr
Executed By: handsOnMod.R \cr
}
