% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyUserModify.R
\name{applyUserModify}
\alias{applyUserModify}
\title{applyUserModify}
\usage{
applyUserModify(
  file.output.list,
  betavalues,
  data_names,
  subdata,
  class_landuse,
  lon_limit
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input and
output of external files.  Created by \code{generateInputList.R}}

\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{subdata}{data.frame input data (subdata)}

\item{class_landuse}{character vector of class_landuses from sparrow_control.R}

\item{lon_limit}{User specified geographic units minimum/maximum longitude limits for
mapping of Residuals and prediction maps}
}
\value{
\code{subdata}  data.frame with all user designated modifications from userModifyData.R
}
\description{
reads \code{userModifyData.R} control file as text, unpacks all variables in the
data1 file and applies user modifications, creates the \code{subdata} object for use in all model
execution statements \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item named.list.R
\item replaceNAs.R
\item unPackList.R} \cr
}
