% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areColors.R
\name{areColors}
\alias{areColors}
\title{areColors}
\usage{
areColors(strColors)
}
\arguments{
\item{strColors}{character vector of supported R color strings}
}
\value{
\code{testCol}  logical TRUE/FALSE indicating whether or not invalid color strings were
found
}
\description{
Tests color strings and indicate whether invalid (not supported by R) color
designations have been found \cr \cr
Executed By: getSpecialSett.R \cr
}
