% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIncremSiteIDs.R
\name{assignIncremSiteIDs}
\alias{assignIncremSiteIDs}
\title{assignIncremSiteIDs}
\usage{
assignIncremSiteIDs(minnum, staid, waterid, tnode, fnode)
}
\arguments{
\item{minnum}{user selected minimum number of reaches between sites}

\item{staid}{reach site IDs (non-zero for reaches with selected monitoring sites)}

\item{waterid}{reach ID number}

\item{tnode}{reach to (downstream) node}

\item{fnode}{reach from (upstream) node}
}
\value{
\code{rchstaid}  integer vector site IDs assigned contiguously to upstream incremental
reaches
}
\description{
Assigns the nearest downstream site ID (staidseq variable based on staid
variable) to reaches in the incremental areas between monitoring sites by climbing network structure
(reaches sorted by HYDSEQ) \cr \cr
Executed By: \itemize{\item selectCalibrationSites.R
\item selectValidationSites.R
\item setNLLSWeights.R} \cr
Executes Routines: sites_incr.for \cr
}
