% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClassLandusePercent.R
\name{calcClassLandusePercent}
\alias{calcClassLandusePercent}
\title{calcClassLandusePercent}
\usage{
calcClassLandusePercent(subdata, class.input.list)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}
}
\value{
\code{percentClassLanduse}  data.frame with calculated percentages by land use
}
\description{
Calculates the land use percentages for the total drainage area for each reach,
based on the user selected land uses given by the class_landuse control setting, followed by a
subsetting of the reaches by the class_landuse_percent control setting. \cr \cr
Executed By: predictSummaryOutCSV.R \cr
Executes Routines: \itemize{\item accumulateIncrArea.R
\item unPackList.R} \cr
}
