% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDemtareaClass.R
\name{calcDemtareaClass}
\alias{calcDemtareaClass}
\title{calcDemtareaClass}
\usage{
calcDemtareaClass(demtarea)
}
\arguments{
\item{demtarea}{total drainage area for each site}
}
\value{
\code{demtarea.class}  a numeric vector indicating decile class labeled with the total
}
\description{
Obtains monitoring station classification (deciles) based on total drainage
area for each site \cr \cr
Executed By: startModelRun.R \cr
}
