% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeadflag.R
\name{calcHeadflag}
\alias{calcHeadflag}
\title{calcHeadflag}
\usage{
calcHeadflag(data1)
}
\arguments{
\item{data1}{input data (data1)}
}
\value{
\code{outdata}  data.frame with waterid and headflag
}
\description{
Calculates the system variable headflag (translated from SAS hydseq code) \cr \cr
Executed By: createVerifyReachAttr.R \cr
Executes Routines: \itemize{\item getVarList.R
\item unPackList.R} \cr
}
