% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIncremLandUse.R
\name{calcIncremLandUse}
\alias{calcIncremLandUse}
\title{calcIncremLandUse}
\usage{
calcIncremLandUse(subdata, class_landuse, idseq,
  minimum_reaches_separating_sites)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}

\item{class_landuse}{character vector of class_landuses from sparrow_control.R}

\item{idseq}{staidseq or vstaidseq, integer vector site IDs assigned contiguously to
upstream incremental reaches}

\item{minimum_reaches_separating_sites}{number indicating the minimum number of reaches
separating sites}
}
\value{
\code{df}  data.frame incremental land-use percentages calculated by class_landuse
}
\description{
Calculates the land-use percentages for the incremental drainage areas of the
monitoring sites for use in decile diagnostics \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: sumIncremAttributes.R \cr
}
