% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTermflag.R
\name{calcTermflag}
\alias{calcTermflag}
\title{calcTermflag}
\usage{
calcTermflag(data1)
}
\arguments{
\item{data1}{input data (data1)}
}
\value{
\code{outdata}  data.frame with waterid and termflag
}
\description{
Calculates the system variable termflag, based on reach navigation variables \cr \cr
Executed By: createVerifyReachAttr.R \cr
Executes Routines: \itemize{\item getVarList.R
\item unPackList.R} \cr
}
