% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBinaryMaps.R
\name{checkBinaryMaps}
\alias{checkBinaryMaps}
\title{checkBinaryMaps}
\usage{
checkBinaryMaps(mapSetting, path_gis, batch_mode)
}
\arguments{
\item{mapSetting}{setting in control file that user sets to use to load the binary map
(\code{lineShapeName}, \code{polyShapeName}, or \code{LineShapeGeo})}

\item{path_gis}{path to users gis data}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{fileLoaded}  logical TRUE/FALSE indicating whether or not file is loaded
}
\description{
Checks if binary mapping objects exist and loads binary files \cr \cr
Executed By: \itemize{\item checkDrainageareaErrors.R
\item diagnosticPlotsNLLS.R
\item diagnosticPlotsValidate.R
\item mapSiteAttributes.R
\item predictMaps.R} \cr
}
