% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClassificationVars.R
\name{checkClassificationVars}
\alias{checkClassificationVars}
\title{checkClassificationVars}
\usage{
checkClassificationVars(subdata, class.input.list, batch_mode)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Checks for missing or zero values in classification variables in section 5 of
the control script. If missing and/or zeros are found, a critical error has been found and
program execution will terminate. \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item errorOccurred.R
\item unPackList.R} \cr
}
