% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData1NavigationVars.R
\name{checkData1NavigationVars}
\alias{checkData1NavigationVars}
\title{checkData1NavigationVars}
\usage{
checkData1NavigationVars(data1, if_reverse_hydseq, batch_mode)
}
\arguments{
\item{data1}{input data (data1)}

\item{if_reverse_hydseq}{yes/no indicating whether hydseq in the DATA1 file needs to be
reversed from sparrow_control}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{data1}  data1 input file as dataframe with large interger values replaced and
hydseq
}
\description{
Checks for excessively large integer values (maximum>1M) for the sparrowNames
\code{waterid}, \code{tnode}, \code{fnode}, \code{hydseq} to avoid array storage problems in R and Fortran
subroutines. Exceedence of the threshold causes a renumbering of the 'waterid' and node numbers. \cr \cr
Executed By: dataInputPrep.R \cr
Executes Routines: errorOccurred.R \cr
}
