% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrainageareaErrors_static.R
\name{checkDrainageareaErrors_static}
\alias{checkDrainageareaErrors_static}
\title{checkDrainageareaErrors}
\usage{
checkDrainageareaErrors_static(
  file.output.list,
  mapping.input.list,
  DAreaFailCheckObj,
  data1,
  batch_mode
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input and
output of external files.  Created by \code{generateInputList.R}}

\item{mapping.input.list}{Named list of sparrow_control settings for mapping: lat_limit,
lon_limit, master_map_list, lineShapeName, lineWaterid,
polyShapeName, ployWaterid, LineShapeGeo, LineShapeGeo, CRStext,
convertShapeToBinary.list, map_siteAttributes.list,
residual_map_breakpoints, site_mapPointScale,
if_verify_demtarea_maps}

\item{DAreaFailCheckObj}{data.frame of all rows of subdata in which the user's original data
for Total Drainage Area vs. Total Drainage Area calculated by RSPARROW differ}

\item{data1}{input data (data1)}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Executes drainage area checks for newly computed areas, based on the
if_verify_demtarea control setting (section 2 of the control script). If any differences are found
between the user's original data for Total Drainage Area vs. Total Drainage Area calculated by
RSPARROW, a plot of user's original data for Total Drainage Area vs. Total Drainage Area calculated
by RSPARROW is output. For the control setting if_verify_demtarea_maps<-"yes", maps are output
of \code{demtarea} and \code{hydseq} for unmatched areas as a ratio of RSPARROW calculated:original. A CSV
file is output of all differences found to ~/estimate/(run_id)_diagnostic_darea_mismatches.csv. \\cr \\cr
Executed By: verifyDemtarea.R \\cr
Executes Routines: \\itemize\{\\item checkBinaryMaps.R
\\item unPackList.R\} \\cr
}
