% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrainageareaMapPrep.R
\name{checkDrainageareaMapPrep}
\alias{checkDrainageareaMapPrep}
\title{checkDrainageareaMapPrep}
\usage{
checkDrainageareaMapPrep(
  file.output.list,
  mapping.input.list,
  DAreaFailCheckObj,
  data1,
  existGeoLines,
  commonvar,
  map.vars.list,
  k
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input and output
of external files.  Created by \code{generateInputList.R}}

\item{mapping.input.list}{Named list of sparrow_control settings for mapping: lat_limit,
lon_limit, master_map_list, lineShapeName, lineWaterid, polyShapeName,
ployWaterid, LineShapeGeo, LineShapeGeo, CRStext,
convertShapeToBinary.list, map_siteAttributes.list,
residual_map_breakpoints, site_mapPointScale, if_verify_demtarea_maps}

\item{DAreaFailCheckObj}{data.frame of all rows of subdata in which the user's original data
for Total Drainage Area vs. Total Drainage Area calculated by RSPARROW
differ}

\item{data1}{input data (data1)}

\item{existGeoLines}{TRUE/FALSE indicating whether the Geolines shape file is present}

\item{commonvar}{string indicating the column to join the map data and the shapefile}

\item{map.vars.list}{character vector indicating which checkDrainageErrors maps to generate}

\item{k}{numeric index for current map}
}
\value{
\code{prepReturns.list} a named list containing \code{dmap} the data to map, \code{break1} vector
of legend breaks, \code{nlty} numeric vector indicating line types,
\code{nlwd} numeric vector of line widths, \code{Mcol}character vector of
unique map colors
}
\description{
Compiles and prepares necessary map data and parameters for checkDrainageErrors
maps \cr \cr
Executed By: checkDrainageareaErrorsChild.Rmd \cr
Executes Routines: \itemize{\item named.list.R
\item unPackList.R} \cr
}
