% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDupVarnames.R
\name{checkDupVarnames}
\alias{checkDupVarnames}
\title{checkDupVarnames}
\usage{
checkDupVarnames(data_names, batch_mode)
}
\arguments{
\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Checks for duplicate \code{sparrowNames} in the dataDictionary.csv file and outputs
list of duplicates as message in console window \cr \cr
Executed By: dataInputPrep.R \cr
}
