% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDynamic.R
\name{checkDynamic}
\alias{checkDynamic}
\title{checkDynamic}
\usage{
checkDynamic(subdata)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}
}
\value{
logical TRUE/FALSE indicating whether the data is dynamic
}
\description{
Determine whether the model is dynamic or static using the
presence of "year" and/or "season" data in subdata \cr \cr
Executed By: \itemize{\item createVerifyReachAttr.R
\item diagnosticPlotsNLLS.Rmd
\item diagnosticPlotsNLLS_timeSeries.Rmd
\item estimate.R
\item estimateNLLSmetrics.R
\item predictScenariosPrep.R
\item readForecast.R
\item startModelRun.R} \cr
}
