% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkingMissingVars.R
\name{checkingMissingVars}
\alias{checkingMissingVars}
\title{checkingMissingVars}
\usage{
checkingMissingVars(checkData, data_names, betavalues, types, allMissing,
  returnData)
}
\arguments{
\item{checkData}{data.frame in which to check for missing values}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{types}{character vector of names of variable lists to check for missing values}

\item{allMissing}{TRUE/FALSE indicating whether to flag 'any missing values' or 'all missing
values'}

\item{returnData}{TRUE/FALSE indicated whether the data checked for missing values should be
returned as part of the \code{missing} list output}
}
\value{
\code{missing} list of number of missing variables \code{k}, names of variables with missing
values \code{datalstMissingdata} and the type of variable list with missing values
}
\description{
check for missing variables in the variable lists specified by \code{types} \cr \cr
Executed By: \itemize{\item checkAnyMissingSubdataVars.R
\item checkMissingData1Vars.R
\item checkMissingSubdataVars.R} \cr
Executes Routines: getVarList.R \cr
}
