% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingData1Vars.R
\name{checkMissingData1Vars}
\alias{checkMissingData1Vars}
\title{checkMissingData1Vars}
\usage{
checkMissingData1Vars(data1, batch_mode)
}
\arguments{
\item{data1}{input data (data1)}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{data1}  data1 input data.frame with missing variables added
}
\description{
Identifies REQUIRED system variables with all missing/zero values in the
\code{data1.csv} file and adds missing variables to the file with all NA values. Outputs a printed warning
for users and a list of missing variables as a message in console. \cr \cr
Executed By: dataInputPrep.R \cr
Executes Routines: \itemize{\item checkingMissingVars.R
\item getVarList.R
\item unPackList.R} \cr
}
