% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingSubdataVars.R
\name{checkMissingSubdataVars}
\alias{checkMissingSubdataVars}
\title{checkMissingSubdataVars}
\usage{
checkMissingSubdataVars(subdata, betavalues, file.output.list, batch_mode)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}

\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Identifies REQUIRED system variables and user selected parameters with all
missing/zero values in \code{subdata} and outputs a list of variables with all missing values as message
in console. \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item checkingMissingVars.R
\item errorOccurred.R
\item getVarList.R
\item unPackList.R} \cr
}
