% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileALL.R
\name{compileALL}
\alias{compileALL}
\title{compileALL}
\usage{
compileALL(input, output, session, path_results, choices)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{path_results}{path to results model subdirectory}

\item{choices}{data.frame output of function createInteractiveChoices.R}
}
\value{
\code{out}  list object of all user input within the namespace of the type of map being
}
\description{
Compiles all user input from the shiny session in the namespace of the map
being generated into a single list object. \cr \cr
Executed By: \itemize{\item goShinyPlot.R
\item shinyMap2.R} \cr
Executes Routines: compileInput.R \cr
}
