% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileInput.R
\name{compileInput}
\alias{compileInput}
\title{compileInput}
\usage{
compileInput(input, output, session)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}
}
\description{
Shiny server function that updates compiles input from namespace indicated in
callModule function call \cr \cr
Executed By: \itemize{\item compileALL.R
\item testCosmetic.R} \cr
}
