% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHotTables.R
\name{convertHotTables}
\alias{convertHotTables}
\title{convertHotTables}
\usage{
convertHotTables(shinyInput)
}
\arguments{
\item{shinyInput}{compiled Shiny input selections}
}
\value{
\code{outshinyInput}  the compiled Shiny Input list with hottables as dataframes and
}
\description{
Converts rhandsontable hottables from Shiny into regular dataframes and
cosmetic mapping settings to regular list objects in the shiny input list \cr \cr
Executed By: \itemize{\item goShinyPlot.R
\item testCosmetic.R
\item testRedTbl.R} \cr
}
