% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyPriorModelFiles.R
\name{copyPriorModelFiles}
\alias{copyPriorModelFiles}
\title{copyPriorModelFiles}
\usage{
copyPriorModelFiles(activeFile, old_run_id, path_master, batch_mode)
}
\arguments{
\item{activeFile}{character string path to sparrow_control.R file at currently top level of
user's results directory}

\item{old_run_id}{character string indicating the archived model subdirectory containing
control files to be copied}

\item{path_master}{character string path to RSPARROW_master directory.  Internally reset to
'RSPARROW_master/R/' subdirectory}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Copies the previously executed model control files into the top level user's
results directory. RSPARROW is terminated after files are copied and the copied control script,
sparrow_control.R, is opened in Rstudio. \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: errorOccurred.R \cr
}
