% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diagnosticPlotList.R
\name{create_diagnosticPlotList}
\alias{create_diagnosticPlotList}
\title{create_diagnosticPlotList}
\usage{
create_diagnosticPlotList()
}
\value{
list of the following diagnostic plots (by plot title)
\itemize{\item ModEstimation_Obs_v_Pred
\item ModEstimation_Box_and_Quantile_Resid
\item ModEstimation_Conditioned_v_Unconditioned_loads
\item ModEstimation_Correlations_ExplanitoryVariables
\item ModEstimation_DrainageArea_DecileClass_Box
\item ModEstimation_Classvar_Decile_Box
\item ModEstimation_LanduseClass_Decile_Box
\item ModEstimation_4panel_Classvar
\item ModSimulation_Obs_v_Pred
\item ModSimulation_Box_and_Quantile_Resid
\item ModSimulation_Correlations_ExplanitoryVariables
\item ModSimulation_DrainageArea_DecileClass_Box
\item ModSimulation_Classvar_Decile_Box
\item ModSimulation_LanduseClass_Decile_Box
\item ModSimulation_4panel_Classvar
\item DiagSensitivity_by_Param
\item DiagSensitivity1\%_by_Param
\item DiagSensitivity1\%_by_Param_logScale
\item CDF_of_Station_Hydrological_Distances
\item CDF_of_Station_Euclidean_Distances
\item Morans_I_by_river_basins
\item Morans_I_by_Class_variable}
}
\description{
Create a list of diagnostic plot functions to be applied to data
at a specific timestep. \cr \cr
Executed By: \itemize{\item diagnosticClassLandChild.Rmd
\item diagnosticClassvarChild.Rmd
\item diagnosticContiguousChild.Rmd
\item diagnosticCorrChild.Rmd
\item diagnosticPlotsNLLS.Rmd
\item diagnosticPlotsNLLS_dyn.R
\item diagnosticPlotsNLLS_dynOut.Rmd
\item diagnosticSensitivity.Rmd
\item diagnosticSensParamChild.Rmd
\item diagnosticSpatialAutoCorr.Rmd} \cr
Executes Routines: \itemize{\item named.list.R
\item unPackList.R
\item diagnosticPlots_4panel_A.R
\item diagnosticPlots_4panel_B.R
\item addMarkerText.R
\item plotlyLayout.R
\item fixDupLatLons.R} \cr
}
