% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataMatrix.R
\name{createDataMatrix}
\alias{createDataMatrix}
\title{createDataMatrix}
\usage{
createDataMatrix(if_mean_adjust_delivery_vars, subdata, SelParmValues,
  betavalues)
}
\arguments{
\item{if_mean_adjust_delivery_vars}{yes/no character string indicating if the delivery
variables are to be mean adjusted from sparrow_control}

\item{subdata}{data.frame input data (subdata)}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{betavalues}{data.frame of model parameters from parameters.csv}
}
\value{
\code{DataMatrix.list}  list containing objects \code{dataNames} (Names of the data matrix
columns),
}
\description{
Creates the R \code{DataMatrix.list} object, containing five data elements that are
used to estimate the model and produce model predictions. \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item getVarList.R
\item named.list.R
\item unPackList.R} \cr
}
