% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirs.R
\name{createDirs}
\alias{createDirs}
\title{createDirs}
\usage{
createDirs(file.output.list, if_userModifyData, batch_mode)
}
\arguments{
\item{if_userModifyData}{yes/no indicating whether or not the userModifyData.R control file
is to be applied}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{complete}  TRUE/FALSE logical indicating whether or not function ran successfully
}
\description{
creates all model subdirectories for file output \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item errorOccurred.R
\item syncVarNames.R
\item unPackList.R} \cr
}
