% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInitialDataDictionary.R
\name{createInitialDataDictionary}
\alias{createInitialDataDictionary}
\title{createInitialDataDictionary}
\usage{
createInitialDataDictionary(file.output.list, input_data_fileName,
  create_initial_parameterControlFiles)
}
\arguments{
\item{input_data_fileName}{name of users data1 file}

\item{create_initial_parameterControlFiles}{yes/no indicating if new parameter files should
be generated based on the dataDictionary.csv control file}
}
\description{
Creates a new dataDictionary.csv file based on the column names of \code{data1.csv},
adds missing, REQUIRED, and FIXED variables with \code{data1UserName=="NA"}, and opens the new
dataDictionary.csv file for the user to edit. RSPARROW execution will be terminated once the
dataDictionary.csv file is created. \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item getVarList.R
\item readData.R
\item unPackList.R} \cr
}
