% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInitialParameterControls.R
\name{createInitialParameterControls}
\alias{createInitialParameterControls}
\title{createInitialParameterControls}
\usage{
createInitialParameterControls(file.output.list, batch_mode)
}
\arguments{
\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Creates parameters.csv file and design_matrix.csv based on the \code{varType} column
of the dataDictionary.csv file. Opens the new parameters.csv and design_matrix.csv files for
edit. RSPARROW execution is terminated. \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item importCSVcontrol.R
\item unPackList.R} \cr
}
