% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInteractiveChoices.R
\name{createInteractiveChoices}
\alias{createInteractiveChoices}
\title{createInteractiveChoices}
\usage{
createInteractiveChoices(SelParmValues, existPredict, subdata, data_names,
  map_uncertainties)
}
\arguments{
\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{existPredict}{logical TRUE/FALSE whether predict.list is available in shiny environment}

\item{subdata}{data.frame input data (subdata)}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{map_uncertainties}{Vector of user selected uncertainty parameters to map, if
uncertainty analysis was not run NA}
}
\value{
\code{Choices}  data.frame of parameter metrics by type ("Load Predictions", "Yield
}
\description{
Generates list of prediction metrics for user to select to plot in RShiny based
on modeled parameters \cr \cr
Executed By: shinyMap2.R \cr
}
