% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMasterDataDictionary.R
\name{createMasterDataDictionary}
\alias{createMasterDataDictionary}
\title{createMasterDataDictionary}
\usage{
createMasterDataDictionary(file.output.list, batch_mode)
}
\arguments{
\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Compiles all archived dataDictionary.csv files in the user's results directory
by \code{run_id} from the earliest date forward. The master_dataDictionary.csv file is saved to the
upper level of the user's results directory. \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item getVarList.R
\item importCSVcontrol.R
\item unPackList.R} \cr
}
