% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRTables.R
\name{createRTables}
\alias{createRTables}
\title{createRTables}
\usage{
createRTables(selectSources, data_names, mapping.input.list)
}
\arguments{
\item{selectSources}{character vector of SOURCE parameters defined as
as.character(JacobResults$Parmnames\link{which(JacobResults$btype=="SOURCE")}) if \code{JacobResults} object is available}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{mapping.input.list}{Named list of sparrow_control settings for mapping: lat_limit,
lon_limit, master_map_list, lineShapeName, lineWaterid,
polyShapeName, ployWaterid, LineShapeGeo, LineShapeGeo, CRStext,
convertShapeToBinary.list, map_siteAttributes.list,
residual_map_breakpoints, site_mapPointScale,
if_verify_demtarea_maps}
}
\value{
\code{out}  list containing all user input tables for the Rshiny app.
}
\description{
Creates user input tables to be converted into hottables in Rshiny. \cr \cr
Executed By: shinyMap2.R \cr
Executes Routines: named.list.R \cr
}
