% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubdataSorted.R
\name{createSubdataSorted}
\alias{createSubdataSorted}
\title{createSubdataSorted}
\usage{
createSubdataSorted(filter_data1_conditions, data1)
}
\arguments{
\item{filter_data1_conditions}{User specified additional DATA1 variables (and conditions) to
be used to filter reaches from sparrow_control}

\item{data1}{input data (data1)}
}
\value{
\code{subdata}  data.frame used in model execution
}
\description{
Creates a subset of \code{data1}, called \code{subdata}, sorted by \code{hydseq} , based on
the application of the control setting 'filter_data1_conditions'. The 'subdata' object is used in
model estimation, prediction, mapping, and other functions.  \cr \cr
Executed By: startModelRun.R \cr
}
