% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVerifyReachAttr.R
\name{createVerifyReachAttr}
\alias{createVerifyReachAttr}
\title{createVerifyReachAttr}
\usage{
createVerifyReachAttr(
  if_verify_demtarea,
  calculate_reach_attribute_list,
  data1,
  file.output.list,
  mapping.input.list,
  batch_mode
)
}
\arguments{
\item{if_verify_demtarea}{specify whether or not to verify demtarea}

\item{calculate_reach_attribute_list}{list of attributes to calculate}

\item{data1}{input data (data1)}

\item{file.output.list}{list of control settings and relative paths used for input and
output of external files.  Created by \code{generateInputList.R}}

\item{mapping.input.list}{Named list of sparrow_control settings for mapping: lat_limit,
lon_limit, master_map_list, lineShapeName, lineWaterid,
polyShapeName, ployWaterid, LineShapeGeo, LineShapeGeo, CRStext,
convertShapeToBinary.list, map_siteAttributes.list,
residual_map_breakpoints, site_mapPointScale,
if_verify_demtarea_maps}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{data1}  data.frame with calculated reach attributes replaced
}
\description{
Calculates REQUIRED system reach attributes for reaches with \code{fnode>0},
\code{tnode>0} and \code{termflag!=3}from \code{calculate_reach_attribute_list} control setting and verifies
\code{demtarea} if \code{if_verify_demtarea<-"yes"} \cr \cr
Executed By: dataInputPrep.R \cr
Executes Routines: \itemize{\item accumulateIncrArea.R
\item calcHeadflag.R
\item calcTermflag.R
\item errorOccurred.R
\item hydseq.R
\item unPackList.R
\item verifyDemtarea.R} \cr
}
