% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInputPrep.R
\name{dataInputPrep}
\alias{dataInputPrep}
\title{dataInputPrep}
\usage{
dataInputPrep(file.output.list, input_data_fileName, if_reverse_hydseq,
  if_verify_demtarea, calculate_reach_attribute_list, mapping.input.list,
  batch_mode)
}
\arguments{
\item{input_data_fileName}{name of users data1 file}

\item{if_reverse_hydseq}{yes/no indicating whether hydseq in the DATA1 file needs to be
reversed from sparrow_control}

\item{if_verify_demtarea}{specify whether or not to verify demtarea}

\item{calculate_reach_attribute_list}{list of attributes to calculate}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Imports data1.csv and dataDictionary.csv files, checks for duplicate
\code{sparrowNames}, replaces \code{data1UserNames} with \code{sparrowNames} in the \code{data1} object, checksreach
navigation variables for large integers and/or all missing values, creates and/or verifies navigation
variables and \code{demtarea}, saves \code{data1} and \code{data_names} (dataDictionary.csv import object)
objects as binary file in top level of results (input_data_fileName)_priorImport, and saves
\code{data_names} and \code{data1} to \code{.GlobalEnv} \cr \cr
Executed By: \itemize{\item batchRun.R
\item executeRSPARROW.R} \cr
Executes Routines: \itemize{\item checkData1NavigationVars.R
\item checkDupVarnames.R
\item checkMissingData1Vars.R
\item createVerifyReachAttr.R
\item read_dataDictionary.R
\item readData.R
\item replaceData1Names.R
\item unPackList.R} \cr
}
