% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFiles.R
\name{deleteFiles}
\alias{deleteFiles}
\title{deleteFiles}
\usage{
deleteFiles(path_results)
}
\arguments{
\item{path_results}{path to results model subdirectory}
}
\description{
Deletes all files from estimate, predict, maps, and scenarios subdirectories.
Subroutine is only called when model estimation is requested via the control settings
if_estimate<-"yes" or if_estimate_simulation<-"yes". \cr \cr
Executed By: executeRSPARROW.R \cr
}
