% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlots_4panel_A.R
\name{diagnosticPlots_4panel_A}
\alias{diagnosticPlots_4panel_A}
\title{diagnosticPlots_4panel_A}
\usage{
diagnosticPlots_4panel_A(
  plotpredict,
  plotObs,
  plotyldpredict,
  plotyldobs,
  sitedata,
  plotResids,
  plotclass,
  plotTitles,
  loadUnits,
  yieldUnits,
  showPlotGrid,
  markerList,
  add_plotlyVars,
  pnch,
  markerCols,
  hline,
  filterClass
)
}
\arguments{
\item{plotpredict}{numeric vector of load prediction values}

\item{plotObs}{numeric vector of load observation values}

\item{plotyldpredict}{numeric vector of yield prediction values}

\item{plotyldobs}{numeric vector of yield observation values}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0 & subdata$calsites==1), ]}. The object contains the dataDictionary
‘sparrowNames’ variables, with records sorted in hydrological
(upstream to downstream) order (see the documentation Chapter
sub-section 5.1.2 for details)}

\item{plotResids}{numeric vector of residuals}

\item{plotclass}{character string indicating which class (selection of classvar
or class_landuse) to filter according to \code{filterClass}}

\item{plotTitles}{character vector of plot titles for 4-panel}

\item{loadUnits}{character string RSPARROW user setting defining units for load}

\item{yieldUnits}{character string RSPARROW user setting defining units for yield}

\item{showPlotGrid}{yes/no setting controlling whether gridlines are displayed}

\item{markerList}{character string defining plotly marker hover text}

\item{add_plotlyVars}{character vector indicating user selected variables to add to plot hover
text}

\item{pnch}{numeric vector of pnch point styles}

\item{markerCols}{vector of hexodecimal color values}

\item{hline}{function to create horizontal red line}

\item{filterClass}{numeric vector of classvar variable being plotted}
}
\value{
4 panel diagnostic plot of "Observed vs Predicted Load","Observed
vs Predicted Yield","Residuals vs Predicted Load","Residuals vs Predicted Yield"
}
\description{
Generates 4 panel diagnostic plots including "Observed vs
Predicted Load","Observed vs Predicted Yield","Residuals
vs Predicted Load","Residuals vs Predicted Yield" \cr \cr
Executed By: create_diagnosticPlotList.R \cr
Executes Routines: \itemize{\item addMarkerText.R
\item plotlyLayout.R } \cr
}
