% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlots_4panel_B.R
\name{diagnosticPlots_4panel_B}
\alias{diagnosticPlots_4panel_B}
\title{diagnosticPlots_4panel_B}
\usage{
diagnosticPlots_4panel_B(
  sitedata,
  plotResids,
  plot.ratio.obs.pred,
  plot.standardResids,
  plotpredict,
  plotTitles,
  loadUnits,
  yieldUnits,
  showPlotGrid,
  markerList,
  add_plotlyVars,
  pnch,
  markerCols,
  hline
)
}
\arguments{
\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0 & subdata$calsites==1), ]}. The object contains the dataDictionary
‘sparrowNames’ variables, with records sorted in hydrological
(upstream to downstream) order (see the documentation Chapter
sub-section 5.1.2 for details)}

\item{plotResids}{numeric vector of residuals}

\item{plot.ratio.obs.pred}{numeric vector ratio of observed vs predicted with
no monitoring adjustment}

\item{plot.standardResids}{numeric vector of standardized residuals}

\item{plotpredict}{numeric vector of load prediction values}

\item{plotTitles}{character vector of plot titles for 4-panel}

\item{loadUnits}{character string RSPARROW user setting defining units for load}

\item{yieldUnits}{character string RSPARROW user setting defining units for yield}

\item{showPlotGrid}{yes/no setting controlling whether gridlines are displayed}

\item{markerList}{character string defining plotly marker hover text}

\item{add_plotlyVars}{character vector indicating user selected variables to add to plot hover
text}

\item{pnch}{numeric vector of pnch point styles}

\item{markerCols}{vector of hexodecimal color values}

\item{hline}{function to create horizontal red line}
}
\value{
4 panel diagnostic plot of "Residuals","Observed / Predicted Ratio",
"Normal Q-Q Plot","Squared Residuals
}
\description{
Generates 4 panel diagnostic plots including "Residuals",
"Observed / Predicted Ratio","Normal Q-Q Plot","Squared Residuals
vs Predicted Load") \cr \cr
Executed By: create_diagnosticPlotList.R \cr
Executes Routines: \itemize{\item addMarkerText.R
\item plotlyLayout.R } \cr
}
