% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlotsNLLS_dyn.R
\name{diagnosticPlotsNLLS_dyn}
\alias{diagnosticPlotsNLLS_dyn}
\title{diagnosticPlotsNLLS_dyn}
\usage{
diagnosticPlotsNLLS_dyn(diagnostic_params)
}
\arguments{
\item{diagnostic_params}{list of all parameters required for dynamic diagnostics.
Includes: \itemize{\item validation - TRUE/FALSE indicating
whether validation should
be run
\item sensitivity - TRUE/FALSE indicating
whether sensitivity analysis
should be run
\item spatialAutoCorr - TRUE/FALSE indicating whether
spatial auto correlation should
be run
\item file.output.list - list of settings for file
output.  created by
generateInputLists()
\item path_diagnosticMapAttrChild - normalized path to
diagnosticMapAttrChild.Rmd
\item path_diagnosticCorrChild - normalized path to
diagnosticCorrChild.Rmd
\item path_diagnosticClassvarChild - normalized path to
diagnosticClassvarChild.Rmd
\item path_diagnosticClassLandChild - normalized path to
diagnosticClassLandChild.Rmd
\item path_diagnosticContiguousChild - normalized path to
diagnosticContiguousChild.Rmd
\item path_diagnosticDiagMapChild - normalized path to
diagnosticDiagMapChild.Rmd
\item path_outputMapsChild - normalized path to
outputMapsChild.Rmd
\item path_outputMaps - normalized path to
outputMaps.Rmd
\item class.input.list - list of class relate settings
created by generateInputLists()
\item sitedata.demtarea.class - numeric vector indicating decile
class labeled with the total
output by calcDemtareaClass()
\item sitedata - data.frame subset of subdata where depvar>0 and
calsites=1
\item sitedata.landuse - data.frame incremental land-use percentages
calculated by class_landuse output by
calcIncremLandUse()
\item estimate.list - output list of estimation results output by
estimate()
\item mapping.input.list - list of mapping settings output by
generateInputLists()
\item Csites.weights.list - regression weights as proportional to
incremental area size output by
setNLLSWeights()
\item Cor.ExplanVars.list - list of explantory correlations
output by correlationMatrix()
\item data_names - data.frame of dataDictionary.csv
\item add_vars - user setting indicating variables to add to
output data.frames
\item batch_mode - yes/no indicating whether batch mode operation is
in effect}}
}
\description{
run diagnostic plots and maps at each selected timestep for
dynamic data \cr \cr
Executed By: \itemize{\item diagnosticPlotsNLLS.R
\item diagnosticPlotsValidate.R
\item diagnosticSensitivity.R
\item diagnosticSpatialAutoCorr.R} \cr
Executes Routines: \itemize{\item unPackList.R
\item create_diagnosticPlotList.R
\item diagnosticPlotsNLLS_dynOut.Rmd
\item diagnosticPlotsNLLS_timeSeries.Rmd} \cr
}
