% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticResidMapExec.R
\name{diagnosticResidMapExec}
\alias{diagnosticResidMapExec}
\title{diagnosticResidMapExec}
\usage{
diagnosticResidMapExec(
  validation,
  existGeoLines,
  residmapTypes,
  mapColumn,
  strTitle,
  dynamic,
  sitedata,
  Resids,
  pResids,
  ratio.obs.pred,
  pratio.obs.pred,
  standardResids,
  path_diagnosticDiagMapChild,
  path_outputMapsChild,
  path_outputMaps,
  map_years,
  map_seasons,
  mapPageGroupBy,
  mapsPerPage,
  enable_plotlyMaps,
  add_plotlyVars,
  path_gis,
  LineShapeGeo,
  GeoLines,
  data_names,
  mapping.input.list,
  batch_mode
)
}
\arguments{
\item{validation}{TRUE/FALSE indicating whether or not validation is to be run}

\item{residmapTypes}{character string indicating what type of residual maps to
generate from the following list: "threshold-above",
"threshold-below","all"}

\item{mapColumn}{character string name of column with residual data}

\item{strTitle}{character string map title}

\item{dynamic}{TRUE/FALSE indicating whether dynamic data is present}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0 & subdata$calsites==1), ]}. The object contains the dataDictionary
‘sparrowNames’ variables, with records sorted in hydrological
(upstream to downstream) order (see the documentation Chapter
sub-section 5.1.2 for details)}

\item{Resids}{numeric vector of residuals}

\item{pResids}{numeric vector of residuals with no monitoring adjustment}

\item{ratio.obs.pred}{numeric vector ratio of observed vs predicted observations}

\item{pratio.obs.pred}{numeric vector of observed vs predicted observations with
no monitoring adjustment}

\item{standardResids}{numeric vector of standardized residuals}

\item{path_diagnosticDiagMapChild}{normalized path to diagnosticDiagMapChild.Rmd}

\item{path_outputMapsChild}{normalized path to outputMapsChild.Rmd}

\item{path_outputMaps}{normalized path to outputMaps.Rmd}

\item{map_years}{RSPARROW dynamic user setting indicating how dynamic mapping
should proceed with regards to the year timestep.  A numeric
vector of years indicates mapping of each year in the vector.
A character string equal to "all" indicates that all years
present in the data1 file should be mapped. A character string
from the following list of approved aggregation functions
c("mean","median","min","max") indicates that the selected
aggregation function should be applied to the "year" attribute
to obtain a seasonal mean, median, min or max.  NA indicates that
the year timestep is not present in data1}

\item{map_seasons}{RSPARROW dynamic user setting indicating how dynamic mapping
should proceed with regards to the season timestep.  A character
vector of seasons indicates mapping of each season in the vector.
Seasons must be from the following list c("winter","spring","summer"
"fall").A character string equal to "all" indicates that all seasons
present in the data1 file should be mapped. A character string
from the following list of approved aggregation functions
c("mean","median","min","max") indicates that the selected
aggregation function should be applied to the "season" attribute
to obtain an annual mean, median, min or max.  NA indicates that
the season timestep is not present in data1}

\item{mapPageGroupBy}{RSPARROW dynamic user setting indicating grouping variable for
maps on a page, select "season","year",or NA for no group}

\item{mapsPerPage}{number between 1 and 4 indicating maps per page}

\item{enable_plotlyMaps}{yes/no indicating whether to enable plotly interactive
displays for maps}

\item{add_plotlyVars}{character vector indicating user selected variables to add to plot hover
text}

\item{path_gis}{path to users gis data}

\item{LineShapeGeo}{character string control settting indicating which binary map file to
load for the Geolines background layer}

\item{GeoLines}{Optional geospatial shape file for overlay of lines on output maps}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{mapping.input.list}{Named list of sparrow_control settings for mapping: lat_limit,
lon_limit, master_map_list, lineShapeName, lineWaterid,
polyShapeName, ployWaterid, LineShapeGeo, LineShapeGeo, CRStext,
convertShapeToBinary.list, map_siteAttributes.list,
residual_map_breakpoints, site_mapPointScale,
if_verify_demtarea_maps}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}

\item{existGeolines}{TRUE/FALSE indicating whether Geolines object exists}
}
\description{
Creates diagnostic Residual maps \cr \cr
Executed By: diagnosticPlotsNLLS.Rmd \cr
Executes Routines: \itemize{\item diagnosticDiagMapChild.Rmd
\item aggDynamicMapdata.R
\item unPackList.R
\item setupDynamicMaps.R} \cr
}
