% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropFunc.R
\name{dropFunc}
\alias{dropFunc}
\title{dropFunc}
\usage{
dropFunc(id, category, choices)
}
\arguments{
\item{id}{Shiny namespace designation}

\item{category}{variable type from \code{unique(choices$category)} in the Rshiny app}

\item{choices}{data.frame output of function createInteractiveChoices.R}
}
\description{
Shiny ui function creates drop down button with SelectAll and ClearAll Buttons
and a multi-select list of variables in the designated category. Only applied in Batch mode. \cr \cr
Executed By: \itemize{\item shinyScenarios.R
\item shinySiteAttr.R
\item streamCatch.R} \cr
}
