% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigensort.R
\name{eigensort}
\alias{eigensort}
\title{eigensort}
\usage{
eigensort(eigs)
}
\arguments{
\item{eigs}{object containing eigenvalues and vectors from the base R \code{eigen()} function}
}
\value{
\code{eigenout}  list object with eigenvalues and vectors sorted in decreasing order
}
\description{
Sorts eigenvectors and eigenvalues in decreasing order. \cr \cr
Executed By: estimateNLLSmetrics.R \cr
}
