% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorOccurred.R
\name{errorOccurred}
\alias{errorOccurred}
\title{errorOccurred}
\usage{
errorOccurred(scriptName, batch_mode)
}
\arguments{
\item{scriptName}{name of functional routine or '.R' control file}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Terminates program if an error occures in any routine. Error message is printed
to the console. \cr \cr
Executed By: \itemize{\item batchRun.R
\item addVars.R
\item checkAnyMissingSubdataVars.R
\item checkClassificationVars.R
\item checkData1NavigationVars.R
\item checkMissingSubdataVars.R
\item copyPriorModelFiles.R
\item createDirs.R
\item createVerifyReachAttr.R
\item estimateNLLSmetrics.R
\item executeRSPARROW.R
\item findControlFiles.R
\item importCSVcontrol.R
\item makePaths.R
\item readParameters.R
\item selectParmValues.R
\item setNLLSWeights.R} \cr
}
