% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFeval.R
\name{estimateFeval}
\alias{estimateFeval}
\title{estimateFeval}
\usage{
estimateFeval(beta0, DataMatrix.list, SelParmValues, Csites.weights.list,
  estimate.input.list, dlvdsgn)
}
\arguments{
\item{beta0}{estimated parameters (no constants)}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}
}
\value{
\code{e} vector of weighted residuals
}
\description{
SPARROW NonLinear Least Squares (NLLS) function for optimizing the model fit to
the observations for the calibration sites, based on use of conditioned predictions of load.
The function accumulates loads in the reach network according to the user's model specification,
making comparisons of the conditioned predictions of  load to the actual loads at monitored
reaches to return a vector of weighted residuals (difference between the actual and predicted
loads). Monitoring load substitution is performed (ifadjust=1).  \cr \cr
Executed By: \itemize{\item estimate.R
\item estimateBootstraps.R
\item estimateFevalNoadj.R
\item estimateNLLSmetrics.R
\item estimateOptimize.R
\item validateFevalNoadj.R} \cr
Executes Routines: \itemize{\item unPackList.R
\item tnoder.for} \cr
}
