% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateOptimize.R
\name{estimateOptimize}
\alias{estimateOptimize}
\title{estimateOptimize}
\usage{
estimateOptimize(file.output.list, SelParmValues, estimate.input.list,
  DataMatrix.list, dlvdsgn)
}
\arguments{
\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}
}
\value{
\code{sparrowEsts} list object contained in estimate.list \code{if_estimate<-'yes'}.  For more
details see documentation Section 5.2.4.4.
}
\description{
Executes the SPARROW NLLS model estimation via calls to the 'nlfb' NLMRT
library optimization function and the 'jacobian' function.  \cr \cr
Executed By: estimate.R \cr
Executes Routines: \itemize{\item estimateFeval.R
\item named.list.R
\item unPackList.R} \cr
}
