% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWeightedErrors.R
\name{estimateWeightedErrors}
\alias{estimateWeightedErrors}
\title{estimateWeightedErrors}
\usage{
estimateWeightedErrors(
  file.output.list,
  xrun_id,
  pre_run_id,
  nreaches,
  calsites
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input
and output of external files.  Created by \code{generateInputList.R}}

\item{xrun_id}{current model run_id used by \code{estimateWeightedErrors.R} executed in
\code{userModifyData.R}}

\item{pre_run_id}{run_id for the prior model residuals used by \code{estimateWeightedErrors.R}
executed in \code{userModifyData.R}}

\item{nreaches}{number of reaches}

\item{calsites}{calibration site indicator (NA or 0 = not selected; 1=selected)}
}
\value{
\code{weight} numeric vector of weights computed as the reciprocal of the
squared residutals (variance) predicted by the power function
}
\description{
Estimates the residual weights for a SPARROW model according to the
"lnload" method, with the weights estimated using a nonlinear power
function regression of the squared log residuals on log transformed
values of the prediction loads.  \cr \cr
Executes Routines: \itemize{\item named.list.R
\item unPackList.R} \cr
}
