% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeRSPARROW.R
\name{executeRSPARROW}
\alias{executeRSPARROW}
\title{executeRSPARROW}
\usage{
executeRSPARROW(settingValues, settingNames, activeFile,
  envir = .GlobalEnv)
}
\arguments{
\item{settingValues}{user input values for all control settings}

\item{settingNames}{names of all control settings}

\item{activeFile}{character string path to sparrow_control.R file at currently top level of
user's results directory}
}
\description{
Runs manageDirVars functions and then proceeds with model excution either in
regular mode or batch mode \cr \cr
Executed By: runRsparrow.R \cr
Executes Routines: \itemize{\item batchRun.R
\item addVars.R
\item copyPriorModelFiles.R
\item createDirs.R
\item createInitialDataDictionary.R
\item createInitialParameterControls.R
\item dataInputPrep.R
\item deleteFiles.R
\item errorOccurred.R
\item findControlFiles.R
\item generateInputLists.R
\item getCharSett.R
\item getNumSett.R
\item getOptionSett.R
\item getShortSett.R
\item getYesNoSett.R
\item isScriptSaved.R
\item makePaths.R
\item named.list.R
\item openDesign.R
\item openParameters.R
\item openVarnames.R
\item outputSettings.R
\item removeObjects.R
\item setMapDefaults.R
\item setupMaps.R
\item startModelRun.R
\item testSettings.R
\item unPackList.R} \cr
}
