% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executionTree.R
\name{executionTree}
\alias{executionTree}
\title{executionTree}
\usage{
executionTree(path_main, startRoutine = "runRsparrow.R",
  includeTypes = c("manageDirVars", "sparrowSetup", "sparrowExec",
  "shiny", "batch", "errorTrap", "lists", "fortran"), includeList = NA,
  excludeList = c("errorOccurred.R", "named.list.R"),
  allOccurances = FALSE, outputType = "data.tree", pruneTree = NA,
  treeLimit = NA)
}
\arguments{
\item{path_main}{character string path to RSPARROW_master directory}

\item{includeTypes}{character string vector indicating which types of functions/routines to
include in executionTree from the function types in RPSARROW_master/inst/tables/funcTypes.csv}

\item{includeList}{character string vector of functions/routines to include in executionTree
even if function/routine not in includeTypes}

\item{excludeList}{character string vector of functions/routines to exclude in executionTree
even if function/routine in includeTypes}

\item{allOccurances}{TRUE/FALSE indicating whether all executions of selected functions
should be output in the data.table of the executionTree.  If `outputType = 'data.tree',
allOccurances will be reset to FALSE}

\item{pruneTree}{number of execution levels to include in the execution tree}

\item{treeLimit}{number of lines to print in data.tree}
}
\value{
\code{traceProgram} data.table of all function/routine executions and line number of
executions after the \code{startRoutine}
}
\description{
Function to trace all function and routine executions from the \code{startRoutine}
forward and either print a data.tree of the executions or output a data.table with executions
and line numbers of executions \cr \cr
}
