% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCodeStr.R
\name{findCodeStr}
\alias{findCodeStr}
\title{findCodeStr}
\usage{
findCodeStr(path_master, str, strType)
}
\arguments{
\item{path_master}{character string path to RSPARROW_master directory.  Internally reset to
'RSPARROW_master/R/' subdirectory}

\item{str}{character string to find in function}

\item{strType}{type of string search, 'routine' indicates that all instances of a routine
(str) being executed will be found, 'args' indicates that all function arguments for the routine
(str) are found, 'setting' indicates that all functions that use the setting (str) as an
argument are found, and 'all' indicates that all instances of the str should be found}
}
\description{
find all instances where a given routine is executed, determine all arguments
for a given routine, find all routines that use a given control setting or functional argument,
or determine all locations within the system a given string occurs \cr \cr
}
