% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findScriptName.R
\name{findScriptName}
\alias{findScriptName}
\title{findScriptName}
\usage{
findScriptName()
}
\value{
\code{path} character string path to current user directory
}
\description{
function to locate the current user directory \cr \cr
Executed By: sparrow_control.R\cr
}
