% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDupLatLons.R
\name{fixDupLatLons}
\alias{fixDupLatLons}
\title{fixDupLatLons}
\usage{
fixDupLatLons(x)
}
\arguments{
\item{x}{numeric vector of latitude or longitude values}
}
\value{
\code{x} latitude or longitude vector with small random increment added to duplicate
latitudes and longitudes
}
\description{
Adds a small random increment of decimal degrees to the latitude and longitude
for monitoring sites with duplicate values. \cr \cr
Executed By: diagnosticSpatialAutoCorr.R \cr
}
