% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInputLists.R
\name{generateInputLists}
\alias{generateInputLists}
\title{generateInputLists}
\usage{
generateInputLists(settings)
}
\arguments{
\item{settings}{named list of all control settings}
}
\description{
creates lists of control settings for group input into functions \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item named.list.R
\item unPackList.R} \cr
}
