% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCharSett.R
\name{getCharSett}
\alias{getCharSett}
\title{getCharSett}
\usage{
getCharSett()
}
\value{
\code{charSettings} list of all character type control settings
}
\description{
list of all character type control settings \cr \cr
Executed By: \itemize{\item executeRSPARROW.R
\item outputSettings.R
\item runRsparrow.R
\item testSettings.R} \cr
}
