% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumSett.R
\name{getNumSett}
\alias{getNumSett}
\title{getNumSett}
\usage{
getNumSett()
}
\value{
\code{numericSettings} list of all numeric type control settings
}
\description{
list of all numeric type control settings \cr \cr
Executed By: \itemize{\item executeRSPARROW.R
\item outputSettings.R
\item runRsparrow.R
\item testCosmetic.R
\item testSettings.R} \cr
}
