% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptionSett.R
\name{getOptionSett}
\alias{getOptionSett}
\title{getOptionSett}
\usage{
getOptionSett()
}
\value{
\code{optionSettings} list of all control settings with specified optional values
}
\description{
list of all control settings with specified optional values \cr \cr
Executed By: \itemize{\item executeRSPARROW.R
\item outputSettings.R
\item runRsparrow.R
\item testSettings.R} \cr
}
