% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShortSett.R
\name{getShortSett}
\alias{getShortSett}
\title{getShortSett}
\usage{
getShortSett()
}
\value{
\code{shortSettings} list of all control settings that should have a length of 1
}
\description{
list of all control settings that should have a length of 1 \cr \cr
Executed By: \itemize{\item executeRSPARROW.R
\item outputSettings.R
\item runRsparrow.R
\item testSettings.R} \cr
}
