% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecialSett.R
\name{getSpecialSett}
\alias{getSpecialSett}
\title{getSpecialSett}
\usage{
getSpecialSett()
}
\value{
\code{specialSettings} list of settings that require additional testing and the test that
should be run
}
\description{
list of settings that require additional testing and the test that should be run \cr \cr
Executed By: \itemize{\item testCosmetic.R
\item testSettings.R} \cr
Executes Routines: areColors.R \cr
}
