% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarList.R
\name{getVarList}
\alias{getVarList}
\title{getVarList}
\usage{
getVarList()
}
\value{
\code{varOUT} list of all fixed and required variables with explanations
}
\description{
list of all fixed and required variables with explanations \cr \cr
Executed By: \itemize{\item addVars.R
\item calcHeadflag.R
\item calcTermflag.R
\item checkAnyMissingSubdataVars.R
\item checkingMissingVars.R
\item checkMissingData1Vars.R
\item checkMissingSubdataVars.R
\item createDataMatrix.R
\item createInitialDataDictionary.R
\item createMasterDataDictionary.R
\item estimateBootstraps.R
\item estimateNLLSmetrics.R
\item hydseq.R
\item hydseqTerm.R
\item predict.R
\item predictBoot.R
\item predictBootsOutCSV.R
\item predictBootstraps.R
\item predictOutCSV.R
\item predictScenariosOutCSV.R
\item predictSensitivity.R
\item read_dataDictionary.R
\item readDesignMatrix.R
\item readParameters.R
\item replaceData1Names.R
\item syncVarNames.R
\item validateMetrics.R} \cr
Executes Routines: named.list.R \cr
}
