% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYesNoSett.R
\name{getYesNoSett}
\alias{getYesNoSett}
\title{getYesNoSett}
\usage{
getYesNoSett()
}
\value{
\code{yesNoSettings} list of all yes/no control settings
}
\description{
list of all yes/no control settings \cr \cr
Executed By: \itemize{\item executeRSPARROW.R
\item outputSettings.R
\item runRsparrow.R
\item testSettings.R} \cr
}
