% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsOnMod.R
\name{handsOnMod}
\alias{handsOnMod}
\title{handsOnMod}
\usage{
handsOnMod(input, output, session, DF)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{DF}{rhandsontable for cosmetic mapping settings in the shiny app}
}
\description{
shiny module to create rshandsontables \cr \cr
Executed By: \itemize{\item shinyMap2.R
\item shinyScenariosMod.R} \cr
Executes Routines: allowRemoveRow.R \cr
}
