% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsOnUI.R
\name{handsOnUI}
\alias{handsOnUI}
\title{handsOnUI}
\usage{
handsOnUI(id, input)
}
\arguments{
\item{id}{Shiny namespace designation}

\item{input}{top level interactive user input in Shiny app}
}
\description{
shiny ui to setup place for rhandsontable \cr \cr
Executed By: \itemize{\item shinyScenarios.R
\item shinySiteAttr.R
\item streamCatch.R} \cr
}
