% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydseq.R
\name{hydseq}
\alias{hydseq}
\title{hydseq}
\usage{
hydseq(indata, calculate_reach_attribute_list)
}
\arguments{
\item{indata}{data.frame containing reach navigation variables and variables to be selected
for incremental area accumulation}

\item{calculate_reach_attribute_list}{list of attributes to calculate}
}
\value{
\code{out} input data file with RSPARROW calculated hydseq values
}
\description{
Creates values for the REQUIRED system variables specified in the
'calculate_reach_attribute_list' control setting, including  'hydseq', 'headflag',  and/or 'demtarea'.  \cr \cr
Executed By: createVerifyReachAttr.R \cr
Executes Routines: \itemize{\item accumulateIncrArea.R
\item getVarList.R
\item unPackList.R} \cr
}
