% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydseqTerm.R
\name{hydseqTerm}
\alias{hydseqTerm}
\title{hydseqTerm}
\usage{
hydseqTerm(subdata, select_targetReachWatersheds)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}

\item{select_targetReachWatersheds}{Indicate the watershed locations where the scenarios
will be applied. For more details see documenation Section 4.4.9.1}
}
\value{
\code{data1} input data file with RSPARROW calculated hydseq values for the reaches
upstream of the designated target reach for source reduction scenarios
}
\description{
flags all reaches upstream of the designated target reaches for execution of
source reduction scenarios \cr \cr
Executed By: predictScenariosPrep.R \cr
Executes Routines: \itemize{\item getVarList.R
\item unPackList.R} \cr
}
