% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPackages.R
\name{installPackages}
\alias{installPackages}
\title{installPackages}
\usage{
installPackages(if_install_packages, path_master)
}
\arguments{
\item{if_install_packages}{yes/no indicating whether to install required R packages to run
RSPARROW. For more details see documentaion Section 4.4.12}

\item{path_master}{character string path to RSPARROW_master directory.  Internally reset to
'RSPARROW_master/R/' subdirectory}
}
\description{
\cr \cr
Executed By:  \cr
}
