% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isScriptSaved.R
\name{isScriptSaved}
\alias{isScriptSaved}
\title{isScriptSaved}
\usage{
isScriptSaved(scriptName, testDir)
}
\arguments{
\item{scriptName}{name of functional routine or '.R' control file}

\item{testDir}{subdirectory in the results directory with the current run_id.  If this
directory exists it will be overwritten, if the user indicates 'yes' to the question in the console}
}
\value{
\code{saved} 1 indicated that the sparrow_control.R file has been saved or 2 indicated
that it has not been saved and execution should be halted
}
\description{
outputs question to the console asking if the current sparrow_control.R file
has been saved and stopping execution until the user indicates that it has been saved \cr \cr
Executed By: executeRSPARROW.R \cr
}
