% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBreaks.R
\name{mapBreaks}
\alias{mapBreaks}
\title{mapBreaks}
\usage{
mapBreaks(vvar, colors)
}
\arguments{
\item{vvar}{mapping variable as vector}

\item{colors}{character vector of colors used in mapping with the number of colors
indicating the number of breaks in the legend}
}
\value{
\code{outBrks} named list of iprob and brks for mapping
}
\description{
Creates mapping color breakpoints.  \cr \cr
Executed By: \itemize{\item mapSiteAttributes.R
\item predictMaps.R} \cr
Executes Routines: named.list.R \cr
}
