% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLoopStr.R
\name{mapLoopStr}
\alias{mapLoopStr}
\title{mapLoopStr}
\usage{
mapLoopStr(mapType, mapLoopInput.list)
}
\arguments{
\item{mapType}{character string indicating type of map being executed ("site",
"stream", "catchment", or "resid")}

\item{mapLoopInput.list}{list of all objects required for a given mapType}
}
\value{
named list multi-panel maps
}
\description{
Creates list of map objects with number of panels equal to
RSPARROW user setting mapsPerPage \cr \cr
Executed By: \itemize{\item goShinyPlot.R
\item outputMaps.Rmd
\item predictMaps.R
\item predictMaps.Rmd } \cr
Executes Routines: \itemize{\item unPackList.R
\item checkBinaryMaps.R
\item predictMaps_single.R
\item mapSiteAttributes.R
\item diagnosticMaps.R} \cr
}
