% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSiteAttributes.R
\name{mapSiteAttributes}
\alias{mapSiteAttributes}
\title{mapSiteAttributes}
\usage{
mapSiteAttributes(input, attr, path_gis, sitedata, LineShapeGeo,
  data_names, Rshiny, mapColumn, mapdata, GeoLines, mapping.input.list,
  strTitle, unitAttr, batch_mode)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{attr}{character string shiny user input of attribute to map in \code{mapSiteAttributes.R}}

\item{path_gis}{path to users gis data}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}

\item{LineShapeGeo}{character string control settting indicating which binary map file to
load for the Geolines background layer}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{Rshiny}{TRUE/FALSE indicating whether routine is being run from the Shiny app}

\item{mapColumn}{character string indicating column of data to be mapped}

\item{mapdata}{input data.frame with lat, long, and column to be mapped}

\item{GeoLines}{Optional geospatial shape file for overlay of lines on output maps}

\item{strTitle}{character string for plot title}

\item{unitAttr}{character string indicating the unit of the attribute being mapped by
\code{mapSiteAttributes.R}}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
function to execute site attribute mapping \cr \cr
Executed By: \itemize{\item interactiveBatchRun.R
\item diagnosticPlotsNLLS.R
\item goShinyPlot.R} \cr
Executes Routines: \itemize{\item checkBinaryMaps.R
\item mapBreaks.R
\item named.list.R
\item replaceNAs.R
\item unPackList.R} \cr
}
