% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCompare.R
\name{modelCompare}
\alias{modelCompare}
\title{modelCompare}
\usage{
modelCompare(
  file.output.list,
  compare_models,
  modelComparison_name,
  if_spatialAutoCorr
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input and
output of external files.  Created by \code{generateInputList.R}}

\item{compare_models}{character string control setting indicated the run_ids of preivously
run model to which the current model is to be compared}

\item{modelComparison_name}{character string control setting that gives the name of the
model comparision being executed}

\item{if_spatialAutoCorr}{yes/no control setting to specifiy if the spatial autocorrelation
diagnostic graphics for Moran's I test are to be output}
}
\description{
function to compare groups of models to the current model using the output from
'estimate/summaryCSV' directory \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: unPackList.R \cr
}
