% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openParameters.R
\name{openParameters}
\alias{openParameters}
\title{openParameters}
\usage{
openParameters(path_user, results_directoryName)
}
\arguments{
\item{path_user}{character string path to RSPARROW user directory containing results, data,
and gis subdirectories}

\item{results_directoryName}{character string indicating the users results subdirectory name}
}
\description{
opens the parameters.csv control file for edit \cr \cr
Executed By: executeRSPARROW.R \cr
}
