% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSettings.R
\name{outputSettings}
\alias{outputSettings}
\title{outputSettings}
\usage{
outputSettings(file.output.list, save)
}
\arguments{
\item{save}{TRUE/FALSE indicating whether control setting values are to be saved to a csv
file}
}
\value{
\code{settings} data.frame of all user control file settings and setting values
}
\description{
outputs archived list of all control settings \cr \cr
Executed By: \itemize{\item controlFileTasksModel.R
\item executeRSPARROW.R
\item predictScenarios.R
\item setupMaps.R
\item startModelRun.R} \cr
Executes Routines: \itemize{\item getCharSett.R
\item getNumSett.R
\item getOptionSett.R
\item getShortSett.R
\item getYesNoSett.R
\item unPackList.R} \cr
}
