% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyLayout.R
\name{plotlyLayout}
\alias{plotlyLayout}
\title{plotlyLayout}
\usage{
plotlyLayout(
  x,
  y,
  log,
  nTicks,
  digits,
  xTitle,
  xZeroLine,
  xLabs = NA,
  xminTick = NA,
  yTitle,
  yZeroLine,
  ymax = NA,
  ymin = NA,
  ymaxTick = NA,
  plotTitle,
  legend,
  showPlotGrid
)
}
\arguments{
\item{x}{data to plot on x-axis}

\item{y}{data to plot on y-axis}

\item{log}{character string indicating which axes are to be plotted in log space}

\item{digits}{number of decimal places to include in axis labels}

\item{xTitle}{character string for x-axis title}

\item{xZeroLine}{TRUE/FALSE indicating whether zeroline should be included on x-axis}

\item{xLabs}{vector of labels for x-axis}

\item{xminTick}{numeric value indicating lowest value on x-axis}

\item{yTitle}{character string for y-axis title}

\item{yZeroLine}{TRUE/FALSE indicating whether zeroline should be included on y-axis}

\item{ymax}{maximum numeric value to include in data plotted on y-axis}

\item{ymin}{minimum numeric value to include in data plotted on y-axis}

\item{ymaxTick}{maximum value for y-axis labels}

\item{plotTitle}{character string for title of plot}

\item{legend}{TRUE/FALSE indicated whether legend should be included}

\item{showPlotGrid}{yes/no setting controlling whether gridlines are displayed}

\item{nticks}{number of tickmarks on axis}
}
\value{
\code{p} plotly plot object with custom layout
}
\description{
initiates plotly plot and sets up layout for plot area and axes \cr \cr
Executed By: \itemize{\item checkDrainageareaErrors.Rmd
\item diagnosticClassLandChild.Rmd
\item diagnosticClassvarChild.Rmd
\item diagnosticContiguousChild.Rmd
\item diagnosticCorrChild.Rmd
\item diagnosticPlotsNLLS.Rmd
\item diagnosticSensitivity.Rmd
\item diagnosticSensParamChild.Rmd
\item diagnosticSpatialAutoCorr.Rmd} \cr
Executes Routines: named.list.R \cr
}
