% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(estimate.list, estimate.input.list, bootcorrection,
  DataMatrix.list, SelParmValues, subdata)
}
\arguments{
\item{bootcorrection}{numeric vector equal to
\code{estimate.list$JacobResults$mean_exp_weighted_error} unless NULL, then reset to 1.0}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{subdata}{data.frame input data (subdata)}
}
\value{
\code{predict.list} archive with all load and yield prediction variables to provide for
the efficient access and use of predictions in subsequent execution of the parametric bootstrap
predictions and uncertainties, mapping, and scenario evaluations.  For more details see
documentation Section 5.3.1.5
}
\description{
Calculates all conditioned and unconditioned model predictions for reaches for
the control setting if_predict<-"yes".  \cr \cr
Executed By: \itemize{\item controlFileTasksModel.R
\item estimate.R} \cr
Executes Routines: \itemize{\item getVarList.R
\item named.list.R
\item unPackList.R
\item deliv_fraction.for
\item mptnoder.for
\item ptnoder.for} \cr
}
