% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaps_single.R
\name{predictMaps_single}
\alias{predictMaps_single}
\title{predictMaps_single}
\usage{
predictMaps_single(
  mapType,
  mapLoopInput.list,
  p,
  plotdata,
  plotPageData,
  titleStr,
  subTitle,
  legendPos,
  legendJus,
  usedColors,
  mapvarname,
  i
)
}
\arguments{
\item{mapType}{character string indicating type of map being executed ("stream"
or"catchment")}

\item{mapLoopInput.list}{list of all objects required for a given mapType}

\item{p}{plot object with base map layer}

\item{plotdata}{data.frame with variables to map}

\item{plotPageData}{data.frame with mapping colors}

\item{titleStr}{character string multi-panel map title}

\item{subTitle}{character string title for each panel in multi-panel map}

\item{legendPos}{character string legend position}

\item{legendJus}{character string legend justification}

\item{usedColors}{vector of colors already present on \code{p}}

\item{mapvarname}{name of column with colors to map in the form MAPCOLORS1,
MAPCOLORS2, etc}

\item{i}{numeric index of color being mapped.}
}
\value{
list including single map with stream or catchment data and vector
of colors present on map
}
\description{
Create a single trace on stream or catchment map for a given
timestep \cr \cr
Executed By: mapLoopStr.R \cr
Executes Routines: \itemize{\item unPackList.R
\item addMarkerText.R
\item named.list.R} \cr
}
