% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSensitivity.R
\name{predictSensitivity}
\alias{predictSensitivity}
\title{predictSensitivity}
\usage{
predictSensitivity(AEstimate, estimate.list, DataMatrix.list,
  SelParmValues, reach_decay_specification, reservoir_decay_specification,
  subdata)
}
\arguments{
\item{AEstimate}{parameter estimates (original or adjusted by 1 percent)}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{reach_decay_specification}{the SAS IML reach decay function code from sparrow_control}

\item{reservoir_decay_specification}{the SAS IML reservoir decay function code from
sparrow_control}

\item{subdata}{data.frame input data (subdata)}
}
\value{
\code{pload_total} nonadjusted total load
}
\description{
Supports the calculation of parameter sensitivities, executed by the
'diagnosticSensitivity' function, by calculating the unconditioned predictions for an individual parameter. \cr \cr
Executed By: diagnosticSensitivity.R \cr
Executes Routines: \itemize{\item getVarList.R
\item unPackList.R
\item ptnoder.for} \cr
}
