% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(file.output.list, input_data_fileName)
}
\arguments{
\item{input_data_fileName}{name of users data1 file}
}
\value{
\code{data1} input data (data1)
}
\description{
Reads the users input data (either .csv or binary format) \cr \cr
Executed By: \itemize{\item createInitialDataDictionary.R
\item dataInputPrep.R} \cr
Executes Routines: unPackList.R \cr
}
