% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDesignMatrix.R
\name{readDesignMatrix}
\alias{readDesignMatrix}
\title{readDesignMatrix}
\usage{
readDesignMatrix(file.output.list, betavalues, batch_mode)
}
\arguments{
\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{dmatrixin} imported object from design_matrix.csv
}
\description{
Reads the land-to-water and source interaction matrix in the
'design_matrix.csv' file.  \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item getVarList.R
\item importCSVcontrol.R
\item unPackList.R} \cr
}
