% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForecast.R
\name{readForecast}
\alias{readForecast}
\title{readForecast}
\usage{
readForecast(
  file.output.list,
  if_mean_adjust_delivery_vars,
  use_sparrowNames,
  subdata,
  SelParmValues,
  forecast_filename,
  data_names,
  batch_mode
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input and
output of external files.  Created by \code{generateInputList.R}}

\item{if_mean_adjust_delivery_vars}{yes/no character string indicating if the delivery
variables are to be mean adjusted from sparrow_control}

\item{use_sparrowNames}{TRUE/FALSE indicating whether the forecast file
contains sparrowNames as column headers (TRUE) or
contains data1UserNames as column headers (FALSE)}

\item{subdata}{data.frame input data (subdata)}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\verb{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) &  ((parmType=="SOURCE" & parmMin>=0) | parmType!="SOURCE")}}

\item{forecast_filename}{character string path to forecast file}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
data.frame of formatted forecast data
}
\description{
Prepare generic external forecast file for
merging with the explanatory data matrix in
the DSS.\cr \cr
Executed By: predictScenariosPrep.R \cr
Executes Routines: \itemize{\item unPackList.R
\item errorOccurred.R
\item checkDynamic.R} \cr
}
