% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeObjects.R
\name{removeObjects}
\alias{removeObjects}
\title{removeObjects}
\usage{
removeObjects(remove.list)
}
\description{
Removes objects from the global environment if the object exists. Warnings are
suppressed. \cr \cr
Executed By: executeRSPARROW.R \cr
}
