% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNAs.R
\name{replaceNAs}
\alias{replaceNAs}
\title{replaceNAs}
\usage{
replaceNAs(listColumns)
}
\arguments{
\item{listColumns}{a named list of variables in which to replace NAs.}
}
\description{
Replaces all NAs with 0's. Output is saved to parent.frame(). \cr \cr
Executed By: \itemize{\item applyUserModify.R
\item mapSiteAttributes.R} \cr
}
