% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSPARROW-package.R
\docType{package}
\name{RSPARROW-package}
\alias{RSPARROW}
\alias{RSPARROW-package}
\title{RSPARROW: RSPARROW: An R system for SPARROW modeling}
\description{
Setup and execution of the USGS SPARROW water-quality model for a user-defined domain.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://code.usgs.gov/water/stats/RSPARROW}
}

}
\keyword{internal}
