% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchShiny.R
\name{runBatchShiny}
\alias{runBatchShiny}
\title{runBatchShiny}
\usage{
runBatchShiny(path_PastResults, path_shinyBrowser = NA)
}
\arguments{
\item{path_PastResults}{character string path to previously generated results including run_id subdirectory}

\item{path_shinyBrowser}{character string path to browser executible, if NA defualt browser will be used}
}
\description{
runs shiny app on previously executed RSPARROW model in batch mode from user
specified browser \cr \cr
Executes Routines: shinyBatch.R \cr
}
\examples{
path_master<-'./RSPARROW_master'
suppressWarnings(remove(list="runRsparrow"))
devtools::load_all(path_master,recompile = FALSE)
runBatchShiny("~/UserDirectory/results/customRunId") #uses default browser
runBatchShiny("~/UserDirectory/results/customRunId",path_shinyBrowser = "C:/Program Files (x86)/Google/Chrome/Application/chrome.exe")
}
