% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{runShiny}
\usage{
runShiny(file.output.list, enable_interactiveMaps, RSPARROW_errorOption)
}
\arguments{
\item{enable_interactiveMaps}{yes/no control setting indicating whether shiny app should be
triggered at the end of the run}

\item{RSPARROW_errorOption}{}
}
\description{
runs shiny app on previously executed RSPARROW model \cr \cr
Executed By:  \cr
Executes Routines: \itemize{\item shinyMap2.R
\item unPackList.R} \cr
}
