% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAll.R
\name{selectAll}
\alias{selectAll}
\title{selectAll}
\usage{
selectAll(input, output, session, category, choices)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{category}{variable type from \code{unique(choices$category)} in the Rshiny app}

\item{choices}{data.frame output of function createInteractiveChoices.R}
}
\description{
Shiny server function updates variable drop down list selecting all variable
choices in selected namespace given by 'category' \cr \cr
Executed By: shinyMap2.R \cr
}
