% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCalibrationSites.R
\name{selectCalibrationSites}
\alias{selectCalibrationSites}
\title{selectCalibrationSites}
\usage{
selectCalibrationSites(subdata, data_names, min.sites.list)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}
}
\value{
\code{Csites.list} list output from \code{selectCalibrationSites.R} modified in
\code{startModelRun.R}
}
\description{
Identifies the calibration sites based on the 'calsites' indicator variable,
filters the sites according to user control setttings (section 3 of control script), and assigns
a unique sequence number to the sites. \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item assignIncremSiteIDs.R
\item named.list.R
\item unPackList.R} \cr
}
