% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDesignMatrix.R
\name{selectDesignMatrix}
\alias{selectDesignMatrix}
\title{selectDesignMatrix}
\usage{
selectDesignMatrix(SelParmValues, betavalues, dmatrixin)
}
\arguments{
\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{dmatrixin}{imported object from design_matrix.csv}
}
\value{
\code{dlvdsgn} design matrix imported from design_matrix.csv for selected parameters
}
\description{
Creates the 'dlvdsgn' matrix for use in model estimation and prediction, based
on the user-selected model parameters.  \cr \cr
Executed By: startModelRun.R \cr
}
