% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectParmValues.R
\name{selectParmValues}
\alias{selectParmValues}
\title{selectParmValues}
\usage{
selectParmValues(df, if_estimate, if_estimate_simulation, batch_mode)
}
\arguments{
\item{df}{betavalues - list of parameters from parameters.csv}

\item{if_estimate}{yes/no indicating whether or not estimation is run}

\item{if_estimate_simulation}{character string setting from sparrow_control.R indicating
whether estimation should be run in simulation mode only.}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{SelParmValues} selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}
}
\description{
Creates the 'SelParmValues' object with the user-selected parameter attributes
and performs consistency checks on the initial, minimum, and maximum values of the parameters.  \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item errorOccurred.R
\item named.list.R} \cr
}
