% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectValidationSites.R
\name{selectValidationSites}
\alias{selectValidationSites}
\title{selectValidationSites}
\usage{
selectValidationSites(iseed, pvalidate, subdata,
  minimum_reaches_separating_sites, data_names)
}
\arguments{
\item{iseed}{User specified initial seed for the bootstraps from sparrow_control}

\item{pvalidate}{numeric control setting indicating a percentage of calibration sites to
select for validation or if equal to 0 indicates that the user defined valsites variable should be
used to select sites for validation. For more details see documentation Section 4.4.6}

\item{subdata}{data.frame input data (subdata)}

\item{minimum_reaches_separating_sites}{number indicating the minimum number of reaches
separating sites}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}
}
\value{
\code{Vsites.list} named list of sites for validation
}
\description{
Identifies the validation sites according to user-specified control settings
(section 6 of the control script).  \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item assignIncremSiteIDs.R
\item named.list.R
\item unPackList.R} \cr
}
