% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMapDefaults.R
\name{setMapDefaults}
\alias{setMapDefaults}
\title{setMapDefaults}
\usage{
setMapDefaults(settings)
}
\arguments{
\item{settings}{named list of all control settings}
}
\description{
Sets default values for any missing required mapping settings and assigns
default value to global environment. \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: unPackList.R \cr
}
