% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDynamicMaps.R
\name{setupDynamicMaps}
\alias{setupDynamicMaps}
\title{setupDynamicMaps}
\usage{
setupDynamicMaps(
  dmapfinal,
  map_years,
  map_seasons,
  mapPageGroupBy,
  mapsPerPage,
  Rshiny,
  enable_plotlyMaps
)
}
\arguments{
\item{dmapfinal}{data.frame containing variables to map}

\item{map_years}{RSPARROW dynamic user setting indicating how dynamic mapping
should proceed with regards to the year timestep.  A numeric
vector of years indicates mapping of each year in the vector.
A character string equal to "all" indicates that all years
present in the data1 file should be mapped. A character string
from the following list of approved aggregation functions
c("mean","median","min","max") indicates that the selected
aggregation function should be applied to the "year" attribute
to obtain a seasonal mean, median, min or max.  NA indicates that
the year timestep is not present in data1}

\item{map_seasons}{RSPARROW dynamic user setting indicating how dynamic mapping
should proceed with regards to the season timestep.  A character
vector of seasons indicates mapping of each season in the vector.
Seasons must be from the following list c("winter","spring","summer"
"fall").A character string equal to "all" indicates that all seasons
present in the data1 file should be mapped. A character string
from the following list of approved aggregation functions
c("mean","median","min","max") indicates that the selected
aggregation function should be applied to the "season" attribute
to obtain an annual mean, median, min or max.  NA indicates that
the season timestep is not present in data1}

\item{mapPageGroupBy}{RSPARROW dynamic user setting indicating grouping variable for
maps on a page, select "season","year",or NA for no group}

\item{mapsPerPage}{number between 1 and 4 indicating maps per page}

\item{Rshiny}{TRUE/FALSE indicating whether routine is being run from the Shiny app}

\item{enable_plotlyMaps}{yes/no indicating whether to enable plotly interactive
displays for maps}
}
\value{
data.frame of all pages of plots grouped according to map_years, map_seasons,
mapPageGroupBy, and mapsPerPage
}
\description{
Create a data.frame of all plots per page according to
RSPARROW user settings map_years, map_seasons,
mapPageGroupBy, and mapsPerPage \cr \cr
Executed By: \itemize{\item checkDrainageareaErrors.R
\item checkDrainageareaMapPrep.R
\item diagnosticPlotsNLLS.Rmd
\item diagnosticResidMapExec.R
\item goShinyPlot.R
\item predictMaps.R} \cr
}
