% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMaps.R
\name{setupMaps}
\alias{setupMaps}
\title{setupMaps}
\usage{
setupMaps(
  file.output.list,
  mapping.input.list,
  batch_mode,
  RSPARROW_errorOption
)
}
\arguments{
\item{file.output.list}{list of control settings and relative paths used for input and
output of external files.  Created by \code{generateInputList.R}}

\item{mapping.input.list}{Named list of sparrow_control settings for mapping: lat_limit,
lon_limit, master_map_list, lineShapeName, lineWaterid,
polyShapeName, ployWaterid, LineShapeGeo, LineShapeGeo, CRStext,
convertShapeToBinary.list, map_siteAttributes.list,
residual_map_breakpoints, site_mapPointScale,
if_verify_demtarea_maps}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}

\item{RSPARROW_errorOption}{yes/no control setting indicating where the RPSARROW_errorOption
should be applied}
}
\description{
A preprocessor to read shape files and store objects \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item batchGeoLines.R
\item batchlineShape.R
\item batchpolyShape.R
\item outputSettings.R
\item unPackList.R} \cr
}
