% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeFunc.R
\name{shapeFunc}
\alias{shapeFunc}
\title{shapeFunc}
\usage{
shapeFunc(id, input)
}
\arguments{
\item{id}{Shiny namespace designation}

\item{input}{top level interactive user input in Shiny app}
}
\description{
Shiny ui function applies Shiny conditional panel to show option to output to
ERSI shapefile if Batch mode is selected \cr \cr
Executed By: shinyMap2.R \cr
}
