% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyErrorTrap.R
\name{shinyErrorTrap}
\alias{shinyErrorTrap}
\title{shinyErrorTrap}
\usage{
shinyErrorTrap(input, path_results, badSettings, errMsg)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{path_results}{path to results model subdirectory}

\item{badSettings}{data.frame of row and column number is rhandsontables with invalid
entries in the shiny app}

\item{errMsg}{character string custom message indicating invalid entries in the shiny app}
}
\value{
\code{errMsg} character string custom message indicating invalid entries in the shiny app
}
\description{
tests for invalid user inputs in the shiny app and pops up a custom error
message, shiny map execution is halted \cr \cr
Executed By: goShinyPlot.R \cr
}
