% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScenariosMod.R
\name{shinyScenariosMod}
\alias{shinyScenariosMod}
\title{shinyScenariosMod}
\usage{
shinyScenariosMod(input, output, session, scenarioRtables, path_results)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{scenarioRtables}{named list of rhandsontables used as initial values for shiny app}

\item{path_results}{path to results model subdirectory}
}
\description{
shiny module for source change scenario mapping \cr \cr
Executed By: shinyMap2.R \cr
Executes Routines: \itemize{\item handsOnMod.R
\item testCosmetic.R
\item validCosmetic.R} \cr
}
