% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRedFunc.R
\name{sourceRedFunc}
\alias{sourceRedFunc}
\title{sourceRedFunc}
\usage{
sourceRedFunc(shinyInput)
}
\arguments{
\item{shinyInput}{compiled Shiny input selections}
}
\value{
\code{out} named list of source change functions based on user input to rhandsontables in
shiny app
}
\description{
create source change functions from rhandsontables in shiny app \cr \cr
Executed By: goShinyPlot.R \cr
}
