% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startEndmodifySubdata.R
\name{startEndmodifySubdata}
\alias{startEndmodifySubdata}
\title{startEndmodifySubdata}
\usage{
startEndmodifySubdata(data_names, class_landuse)
}
\arguments{
\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{class_landuse}{character vector of class_landuses from sparrow_control.R}
}
\value{
\code{subdata} data.frame input data (subdata)
}
\description{
if userModifyData.R is not run, this function creates the subdata object and
stops execution if any missing class_landuse found \cr \cr
Executed By: startModelRun.R \cr
}
