% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startModelRun.R
\name{startModelRun}
\alias{startModelRun}
\title{startModelRun}
\usage{
startModelRun(file.output.list, if_estimate, if_estimate_simulation,
  if_boot_estimate, if_boot_predict, enable_interactiveMaps,
  filter_data1_conditions, data1, if_userModifyData, data_names,
  class.input.list, min.sites.list, if_validate, iseed, pvalidate,
  mapping.input.list, estimate.input.list, if_predict, biters,
  scenario.input.list, compare_models, modelComparison_name,
  if_spatialAutoCorr, add_vars, batch_mode, RSPARROW_errorOption)
}
\arguments{
\item{if_estimate}{yes/no indicating whether or not estimation is run}

\item{if_estimate_simulation}{character string setting from sparrow_control.R indicating
whether estimation should be run in simulation mode only.}

\item{if_boot_estimate}{yes/no control setting to specify if parametric bootstrap estimation
(Monte Carlo) is to be executed}

\item{if_boot_predict}{yes/no control setting to specify if bootstrap predictions (mean, SE,
confidence intervals) are to be executed}

\item{enable_interactiveMaps}{yes/no control setting indicating whether shiny app should be
triggered at the end of the run}

\item{filter_data1_conditions}{User specified additional DATA1 variables (and conditions) to
be used to filter reaches from sparrow_control}

\item{data1}{input data (data1)}

\item{if_userModifyData}{yes/no indicating whether or not the userModifyData.R control file
is to be applied}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{if_validate}{yes/no indicating whether or not validation is run}

\item{iseed}{User specified initial seed for the bootstraps from sparrow_control}

\item{pvalidate}{numeric control setting indicating a percentage of calibration sites to
select for validation or if equal to 0 indicates that the user defined valsites variable should be
used to select sites for validation. For more details see documentation Section 4.4.6}

\item{if_predict}{yes/no indicating whether or not prediction is run}

\item{biters}{User specified number of parametric bootstrap iterations from sparrow_control}

\item{compare_models}{character string control setting indicated the run_ids of preivously
run model to which the current model is to be compared}

\item{modelComparison_name}{character string control setting that gives the name of the
model comparision being executed}

\item{if_spatialAutoCorr}{yes/no control setting to specifiy if the spatial autocorrelation
diagnostic graphics for Moran's I test are to be output}

\item{add_vars}{additional variables specified by the setting \code{add_vars} to be included in
prediction, yield, and residuals csv and shape files}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}

\item{RSPARROW_errorOption}{}
}
\description{
executes sparrowSetup and sparrowExecution functions \cr \cr
Executed By: \itemize{\item batchRun.R
\item executeRSPARROW.R} \cr
Executes Routines: \itemize{\item applyUserModify.R
\item calcDemtareaClass.R
\item calcIncremLandUse.R
\item checkAnyMissingSubdataVars.R
\item checkClassificationVars.R
\item checkMissingSubdataVars.R
\item controlFileTasksModel.R
\item correlationMatrix.R
\item createDataMatrix.R
\item createMasterDataDictionary.R
\item createSubdataSorted.R
\item findMinMaxLatLon.R
\item modelCompare.R
\item named.list.R
\item outputSettings.R
\item readDesignMatrix.R
\item readParameters.R
\item selectCalibrationSites.R
\item selectDesignMatrix.R
\item selectParmValues.R
\item selectValidationSites.R
\item setNLLSWeights.R
\item shinyMap2.R
\item startEndmodifySubdata.R
\item unPackList.R} \cr
}
