% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rsparrow.R
\name{summary.rsparrow}
\alias{summary.rsparrow}
\title{Summarize SPARROW Model Results}
\usage{
\method{summary}{rsparrow}(object, ...)
}
\arguments{
\item{object}{An object of class "rsparrow".}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class "summary.rsparrow".
}
\description{
Produces a summary of SPARROW model estimation results.
}
\examples{
\dontrun{
model <- rsparrow_model("~/my_model/")
summary(model)
}
}
