% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncVarNames.R
\name{syncVarNames}
\alias{syncVarNames}
\title{syncVarNames}
\usage{
syncVarNames(file.output.list, batch_mode)
}
\arguments{
\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Updates the 'dataDictionary.csv' varType column with the user designated
parmTypes from the 'parameters.csv' file. \cr \cr
Executed By: createDirs.R \cr
Executes Routines: \itemize{\item getVarList.R
\item importCSVcontrol.R
\item unPackList.R} \cr
}
