% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_addPlotlyvars.R
\name{test_addPlotlyvars}
\alias{test_addPlotlyvars}
\title{test_addPlotlyvars}
\usage{
test_addPlotlyvars(add_plotlyVars, subdata, groupVar, maxUnique)
}
\arguments{
\item{add_plotlyVars}{character vector indicating user selected variables to add to plot hover
text}

\item{subdata}{data.frame input data (subdata)}

\item{groupVar}{character string indicating column in subdata
used in grouping (season or year)}

\item{maxUnique}{numeric required length of the grouped add_plotlyVars}
}
\value{
character vector of all add_plotlyVars that can be grouped by
groupVar and a message stating any add_plotlyVars that were
removed due to not being unique within the group.
}
\description{
Test whether variables in RSPARROW user setting
\code{add_plotlyVars} can be aggregated according to
\code{map_years} and/or \code{map_seasons}\cr \cr
Executed By: aggDynamicMapdata.R \cr
}
