% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCosmetic.R
\name{testCosmetic}
\alias{testCosmetic}
\title{testCosmetic}
\usage{
testCosmetic(input, output, session, DF, mapType)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{DF}{rhandsontable for cosmetic mapping settings in the shiny app}

\item{mapType}{character string indicating type of map (prediction, siteAttr, scenario)}
}
\value{
\code{out} named list with bad cosmetic mapping settings and rownumbers for invalid user
inputs to map settings tables in the shiny app
}
\description{
test for invalid user inputs in map settings rhandsontables in shiny app \cr \cr
Executed By: \itemize{\item shinyMap2.R
\item shinyScenariosMod.R} \cr
Executes Routines: \itemize{\item compileInput.R
\item convertHotTables.R
\item getNumSett.R
\item getSpecialSett.R
\item named.list.R
\item unPackList.R} \cr
}
