% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRedTbl.R
\name{testRedTbl}
\alias{testRedTbl}
\title{testRedTbl}
\usage{
testRedTbl(input, output, session, DF)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{DF}{rhandsontable for cosmetic mapping settings in the shiny app}
}
\value{
\code{out} named list with bad source change scenario settings and rownumbers for invalid
user inputs to rhandsontables in the shiny app
}
\description{
test for invalid user inputs in source reduction scenario rhandsontables in
shiny app \cr \cr
Executed By: shinyMap2.R \cr
Executes Routines: \itemize{\item convertHotTables.R
\item named.list.R} \cr
}
