% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSettings.R
\name{testSettings}
\alias{testSettings}
\title{testSettings}
\usage{
testSettings(settings, saved)
}
\arguments{
\item{settings}{named list of all control settings}

\item{saved}{binary value indicated whether the user has saved all changes to the
\code{sparrow_control.R} file}
}
\description{
test for invalid user inputs in shiny app \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item getCharSett.R
\item getNumSett.R
\item getOptionSett.R
\item getShortSett.R
\item getSpecialSett.R
\item getYesNoSett.R
\item unPackList.R} \cr
}
