% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackList.R
\name{unPackList}
\alias{unPackList}
\title{unPackList}
\usage{
unPackList(lists, parentObj, env = parent.frame())
}
\arguments{
\item{lists}{list of variables to unpack into the \code{parent.frame()}}

\item{parentObj}{list of parent objects that \code{lists} are part of, if NA no parent object}
}
\description{
creates individual variables from list objects and places them in the
\code{parent.frame()} environment \cr \cr
Executed By: \itemize{\item batchMaps.R
\item batchRun.R
\item interactiveBatchRun.R
\item addVars.R
\item applyUserModify.R
\item calcClassLandusePercent.R
\item calcHeadflag.R
\item calcTermflag.R
\item checkAnyMissingSubdataVars.R
\item checkClassificationVars.R
\item checkDrainageareaErrors.R
\item checkMissingData1Vars.R
\item checkMissingSubdataVars.R
\item controlFileTasksModel.R
\item correlationMatrix.R
\item createDataMatrix.R
\item createDirs.R
\item createInitialDataDictionary.R
\item createInitialParameterControls.R
\item createMasterDataDictionary.R
\item createVerifyReachAttr.R
\item dataInputPrep.R
\item diagnosticMaps.R
\item diagnosticPlotsNLLS.R
\item diagnosticPlotsValidate.R
\item diagnosticSensitivity.R
\item diagnosticSpatialAutoCorr.R
\item estimate.R
\item estimateBootstraps.R
\item estimateFeval.R
\item estimateFevalNoadj.R
\item estimateNLLSmetrics.R
\item estimateNLLStable.R
\item estimateOptimize.R
\item estimateWeightedErrors.R
\item executeRSPARROW.R
\item generateInputLists.R
\item goShinyPlot.R
\item hydseq.R
\item hydseqTerm.R
\item importCSVcontrol.R
\item mapSiteAttributes.R
\item modelCompare.R
\item outputSettings.R
\item predict.R
\item predictBoot.R
\item predictBootsOutCSV.R
\item predictBootstraps.R
\item predictMaps.R
\item predictOutCSV.R
\item predictScenarios.R
\item predictScenariosOutCSV.R
\item predictScenariosPrep.R
\item predictSensitivity.R
\item predictSummaryOutCSV.R
\item read_dataDictionary.R
\item readData.R
\item readDesignMatrix.R
\item readParameters.R
\item runShiny.R
\item selectCalibrationSites.R
\item selectValidationSites.R
\item setMapDefaults.R
\item setNLLSWeights.R
\item setupMaps.R
\item shinyMap2.R
\item startModelRun.R
\item syncVarNames.R
\item testCosmetic.R
\item testSettings.R
\item validateMetrics.R
\item verifyDemtarea.R} \cr
}
