% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateVariable.R
\name{updateVariable}
\alias{updateVariable}
\title{updateVariable}
\usage{
updateVariable(input, output, session, choices, mapType)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{choices}{data.frame output of function createInteractiveChoices.R}

\item{mapType}{character string indicating type of map (prediction, siteAttr, scenario)}
}
\description{
Shiny server function updates variable drop down list according to the user's
selection of mapping variable type and outputs the definition of the variable in Shiny namespace
= 'nsStreamCatch' \cr \cr
Executed By: shinyMap2.R \cr
}
