% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMetrics.R
\name{validateMetrics}
\alias{validateMetrics}
\title{validateMetrics}
\usage{
validateMetrics(classvar, estimate.list, dlvdsgn, Vsites.list, yieldFactor,
  SelParmValues, subdata, vsitedata, DataMatrix.list)
}
\arguments{
\item{classvar}{character vector of user specified spatially contiguous discrete
classification variables from sparrow_control.  First element is reach classification variable.}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}

\item{yieldFactor}{numeric control setting to specify the yield conversion factor, computed
as \code{Yield = load / demtarea * yieldFactor}}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{subdata}{data.frame input data (subdata)}

\item{vsitedata}{sitedata for validation. Calculated by \code{subdata[(subdata$vdepvar > 0), ]}}
}
\value{
\code{validate.metrics.list} named list including the vANOVA.list and vMdiagnostics.list
}
\description{
Computes all model performance and other summary metrics and diagnostics for
the estimated model for the validation sites, as reported in the ~/estimate/(run_id)_summary.txt
file.  \cr \cr
Executed By: estimate.R \cr
Executes Routines: \itemize{\item getVarList.R
\item named.list.R
\item unPackList.R
\item validateFevalNoadj.R} \cr
}
