% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validCosmetic.R
\name{validCosmetic}
\alias{validCosmetic}
\title{validCosmetic}
\usage{
validCosmetic(input, output, session, DF, rowNum)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{DF}{rhandsontable for cosmetic mapping settings in the shiny app}

\item{rowNum}{rownumbers of invalid settings in the rhandsontables in the shiny app}
}
\description{
highlights invalid user inputs in the map settings rhandsontables in the shiny
app \cr \cr
Executed By: \itemize{\item shinyMap2.R
\item shinyScenariosMod.R} \cr
}
