% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validSetting.R
\name{validSetting}
\alias{validSetting}
\title{validSetting}
\usage{
validSetting(input, output, session, inputIdstr, labelstr)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{output}{shiny list object for session output}

\item{session}{active shiny session}

\item{inputIdstr}{character string naming user inputs in the shiny app}

\item{labelstr}{character string giving the labels associated user inputs in the shiny app}
}
\description{
\cr \cr
Executed By:  \cr
}
