% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.rsparrow.R
\name{vcov.rsparrow}
\alias{vcov.rsparrow}
\title{Extract SPARROW Model Variance-Covariance Matrix}
\usage{
\method{vcov}{rsparrow}(object, ...)
}
\arguments{
\item{object}{An object of class "rsparrow".}

\item{...}{Additional arguments (currently unused).}
}
\value{
A square numeric matrix.
}
\description{
Extracts the variance-covariance matrix from a fitted SPARROW model.
}
\examples{
\dontrun{
model <- rsparrow_model("~/my_model/")
vcov(model)
}
}
