% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyDemtarea.R
\name{verifyDemtarea}
\alias{verifyDemtarea}
\title{verifyDemtarea}
\usage{
verifyDemtarea(if_verify_demtarea, data1, compareData, file.output.list,
  mapping.input.list, batch_mode)
}
\arguments{
\item{if_verify_demtarea}{specify whether or not to verify demtarea}

\item{data1}{input data (data1)}

\item{compareData}{data.frame of reach verification attributes to compare with the users
data1 file}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Performs a consistency check on values for the total drainage area system
variable, 'demtarea', based on the 'if_verify_demtarea' control setting (section 2 of the control
script). The check compares values of the newly-created 'demtarea' (using the 'hydseq.R'
function) with user-supplied values of 'demtarea', and outputs the results to CSV and PDF files. \cr \cr
Executed By: createVerifyReachAttr.R \cr
Executes Routines: \itemize{\item checkDrainageareaErrors.R
\item unPackList.R} \cr
}
