% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlFileTasksModel.R
\name{controlFileTasksModel}
\alias{controlFileTasksModel}
\title{controlFileTasksModel}
\usage{
controlFileTasksModel(file.output.list, SelParmValues, betavalues,
  Csites.weights.list, subdata, data_names, DataMatrix.list, sitedata,
  Vsites.list, vsitedata, numsites, class.input.list, Cor.ExplanVars.list,
  minimum_reaches_separating_sites, if_estimate, if_estimate_simulation,
  dlvdsgn, estimate.input.list, if_predict, if_validate, sitedata.landuse,
  vsitedata.landuse, sitedata.demtarea.class, vsitedata.demtarea.class,
  mapping.input.list, iseed, biters, scenario.input.list, batch_mode,
  RSPARROW_errorOption)
}
\arguments{
\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{subdata}{data.frame input data (subdata)}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}

\item{vsitedata}{sitedata for validation. Calculated by \code{subdata[(subdata$vdepvar > 0), ]}}

\item{numsites}{number of sites selected for calibration}

\item{minimum_reaches_separating_sites}{number indicating the minimum number of reaches
separating sites}

\item{if_estimate}{yes/no indicating whether or not estimation is run}

\item{if_estimate_simulation}{character string setting from sparrow_control.R indicating
whether estimation should be run in simulation mode only.}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}

\item{if_predict}{yes/no indicating whether or not prediction is run}

\item{if_validate}{yes/no indicating whether or not validation is run}

\item{iseed}{User specified initial seed for the bootstraps from sparrow_control}

\item{biters}{User specified number of parametric bootstrap iterations from sparrow_control}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}

\item{RSPARROW_errorOption}{}
}
\value{
\code{runTimes}  named list of run times for bootstrap estimation, bootstrap prediction,
mapping, and estimate.list
}
\description{
Executes all model tasks after checks have been performed, including execution
of model estimation, diagnostics, prediction, scenarios, and mapping \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item batchMaps.R
\item diagnosticSpatialAutoCorr.R
\item estimate.R
\item estimateBootstraps.R
\item named.list.R
\item outputSettings.R
\item predict.R
\item predictBootsOutCSV.R
\item predictBootstraps.R
\item predictOutCSV.R
\item predictScenarios.R
\item unPackList.R} \cr
}
