% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{correlationMatrix}
\alias{correlationMatrix}
\title{correlationMatrix}
\usage{
correlationMatrix(file.output.list, SelParmValues, subdata)
}
\arguments{
\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{subdata}{data.frame input data (subdata)}
}
\value{
\code{outchar}  named list Cor.ExplanVars.list
}
\description{
Calculates summary metrics and bivariate correlations between user-selected
variables in \code{subdata} for incremental areas of monitoring sites and for all reaches. Outputs
(run_id)_explvars_correlations.txt and (run_id)_explvars_correlations.pdf files. \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: \itemize{\item named.list.R
\item sumIncremAttributes.R
\item unPackList.R} \cr
}
