% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticMaps.R
\name{diagnosticMaps}
\alias{diagnosticMaps}
\title{diagnosticMaps}
\usage{
diagnosticMaps(mapColumn, mapdata, GeoLines, map.list, strTitle,
  mapping.input.list)
}
\arguments{
\item{mapColumn}{character string indicating column of data to be mapped}

\item{mapdata}{input data.frame with lat, long, and column to be mapped}

\item{GeoLines}{Optional geospatial shape file for overlay of lines on output maps}

\item{strTitle}{character string for plot title}
}
\description{
Creates diagnostic maps of residuals and site attributes and saves them to
~/estimate/(run_id)_diagnostic_plots.pdf. \cr \cr
Executed By: \itemize{\item diagnosticPlotsNLLS.R
\item diagnosticPlotsValidate.R} \cr
Executes Routines: unPackList.R \cr
}
