% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlotsNLLS.R
\name{diagnosticPlotsNLLS}
\alias{diagnosticPlotsNLLS}
\title{diagnosticPlotsNLLS}
\usage{
diagnosticPlotsNLLS(file.output.list, class.input.list,
  sitedata.demtarea.class, sitedata, sitedata.landuse, estimate.list,
  mapping.input.list, Csites.weights.list, Cor.ExplanVars.list, data_names,
  add_vars, batch_mode)
}
\arguments{
\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{add_vars}{additional variables specified by the setting \code{add_vars} to be included in
prediction, yield, and residuals csv and shape files}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Creates diagnostic plots and maps output to
~/estimate/(run_id)_diagnostic_plots.pdf, and saves residual maps as shape files. \cr \cr
Executed By: estimate.R \cr
Executes Routines: \itemize{\item checkBinaryMaps.R
\item diagnosticMaps.R
\item mapSiteAttributes.R
\item unPackList.R} \cr
}
