% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlotsValidate.R
\name{diagnosticPlotsValidate}
\alias{diagnosticPlotsValidate}
\title{diagnosticPlotsValidate}
\usage{
diagnosticPlotsValidate(file.output.list, class.input.list,
  vsitedata.demtarea.class, vsitedata, vsitedata.landuse, estimate.list,
  mapping.input.list, add_vars, batch_mode)
}
\arguments{
\item{vsitedata}{sitedata for validation. Calculated by \code{subdata[(subdata$vdepvar > 0), ]}}

\item{add_vars}{additional variables specified by the setting \code{add_vars} to be included in
prediction, yield, and residuals csv and shape files}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\description{
Creates diagnostic plots and maps for validation sites output to
~/estimate/(run_id)_validation_plots.pdf, and saves residual maps as shape files. \cr \cr
Executed By: estimate.R \cr
Executes Routines: \itemize{\item checkBinaryMaps.R
\item diagnosticMaps.R
\item unPackList.R} \cr
}
