% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticSpatialAutoCorr.R
\name{diagnosticSpatialAutoCorr}
\alias{diagnosticSpatialAutoCorr}
\title{diagnosticSpatialAutoCorr}
\usage{
diagnosticSpatialAutoCorr(file.output.list, classvar, sitedata, numsites,
  estimate.list, estimate.input.list, subdata)
}
\arguments{
\item{classvar}{character vector of user specified spatially contiguous discrete
classification variables from sparrow_control.  First element is reach classification variable.}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}

\item{numsites}{number of sites selected for calibration}

\item{subdata}{data.frame input data (subdata)}
}
\description{
Generates diagnostics for spatial dependence (autocorrelation) in the model
residuals. SiteIDs are determined from hydseq sorted file to ensure consistency in hydrologic
distance and other ID ordering across programs. Outputs plots to
~/estimate/(run_id)_diagnostic_spatialautocor.pdf. Outputs Morans I stats to ~/estimate/(run_id)_diagnostic_spatialautocor.txt.
Outputs Morans I stats to ~/estimate/summaryCSV/(run_id)_EuclideanMoransI.csv \cr \cr
Executed By: controlFileTasksModel.R \cr
Executes Routines: \itemize{\item fixDupLatLons.R
\item unPackList.R} \cr
}
