% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(if_estimate, if_predict, file.output.list, class.input.list,
  dlvdsgn, estimate.input.list, minimum_reaches_separating_sites,
  DataMatrix.list, SelParmValues, Csites.weights.list, Csites.list,
  sitedata, numsites, if_validate, Vsites.list, vsitedata, subdata,
  Cor.ExplanVars.list, sitedata.landuse, vsitedata.landuse,
  sitedata.demtarea.class, vsitedata.demtarea.class, mapping.input.list,
  betavalues, if_estimate_simulation, batch_mode)
}
\arguments{
\item{if_estimate}{yes/no indicating whether or not estimation is run}

\item{if_predict}{yes/no indicating whether or not prediction is run}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}

\item{minimum_reaches_separating_sites}{number indicating the minimum number of reaches
separating sites}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}

\item{numsites}{number of sites selected for calibration}

\item{if_validate}{yes/no indicating whether or not validation is run}

\item{vsitedata}{sitedata for validation. Calculated by \code{subdata[(subdata$vdepvar > 0), ]}}

\item{subdata}{data.frame input data (subdata)}

\item{betavalues}{data.frame of model parameters from parameters.csv}

\item{if_estimate_simulation}{character string setting from sparrow_control.R indicating
whether estimation should be run in simulation mode only.}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}
}
\value{
\code{estimate.list} named list of summary metrics and diagnostic output. For more
details see documentation section 5.2.4
}
\description{
Executes all model tasks related to model estimation, diagnostics plotting and
mapping, parameter sensitivities, validation diagnostics, and the output of tabular summary metrics. \cr \cr
Executed By: controlFileTasksModel.R \cr
Executes Routines: \itemize{\item diagnosticPlotsNLLS.R
\item diagnosticPlotsValidate.R
\item diagnosticSensitivity.R
\item errorOccurred.R
\item estimateFeval.R
\item estimateFevalNoadj.R
\item estimateNLLSmetrics.R
\item estimateNLLStable.R
\item estimateOptimize.R
\item named.list.R
\item predict.R
\item predictSummaryOutCSV.R
\item unPackList.R
\item validateMetrics.R} \cr
}
