% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateBootstraps.R
\name{estimateBootstraps}
\alias{estimateBootstraps}
\title{estimateBootstraps}
\usage{
estimateBootstraps(iseed, biters, estimate.list, DataMatrix.list,
  SelParmValues, Csites.weights.list, estimate.input.list, dlvdsgn,
  file.output.list)
}
\arguments{
\item{iseed}{User specified initial seed for the bootstraps from sparrow_control}

\item{biters}{User specified number of parametric bootstrap iterations from sparrow_control}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}
}
\value{
\code{BootResults} data archive using the control setting \code{if_boot_estimate <- "yes"} for
use in subsequent execution of parametric bootstrap predictions. For more details see
documenation Section 5.2.6.2.
}
\description{
Executes all model tasks related to obtaining model estimation uncertainties
using parametric bootstrap methods, for the control setting if_boot_estimate<-"yes". \cr \cr
Executed By: controlFileTasksModel.R \cr
Executes Routines: \itemize{\item estimateFeval.R
\item getVarList.R
\item named.list.R
\item unPackList.R} \cr
}
