% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFevalNoadj.R
\name{estimateFevalNoadj}
\alias{estimateFevalNoadj}
\title{estimateFevalNoadj}
\usage{
estimateFevalNoadj(beta0, DataMatrix.list, SelParmValues,
  Csites.weights.list, estimate.input.list, dlvdsgn)
}
\arguments{
\item{beta0}{estimated parameters (no constants)}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{dlvdsgn}{design matrix imported from design_matrix.csv}
}
\value{
\code{e} vector of weighted residuals
}
\description{
Accumulates loads in the reach network according to the user's model
specification, making comparisons of the unconditioned predictions of  load to the actual loads for
monitored (calibration site) reaches to return a vector of weighted residuals (difference between
the actual and predicted loads). No monitoring load substitution is performed (ifadjust=0).  \cr \cr
Executed By: \itemize{\item estimate.R
\item estimateNLLSmetrics.R} \cr
Executes Routines: \itemize{\item estimateFeval.R
\item unPackList.R
\item tnoder.for} \cr
}
