% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNLLStable.R
\name{estimateNLLStable}
\alias{estimateNLLStable}
\title{estimateNLLStable}
\usage{
estimateNLLStable(file.output.list, if_estimate, if_estimate_simulation,
  ifHess, if_sparrowEsts, classvar, sitedata, numsites, ANOVA.list,
  JacobResults, HesResults, sparrowEsts, Mdiagnostics.list,
  Cor.ExplanVars.list, if_validate, vANOVA.list, vMdiagnostics.list,
  betavalues, Csites.weights.list)
}
\arguments{
\item{if_estimate}{yes/no indicating whether or not estimation is run}

\item{if_estimate_simulation}{character string setting from sparrow_control.R indicating
whether estimation should be run in simulation mode only.}

\item{ifHess}{yes/no indicating whether second-order Hessian standard errors should be
computed}

\item{if_sparrowEsts}{value of 1 \code{if_estimate_simulation<-'no'} indicating calculation of
Jacobian diagnostics and 0 \code{if_estimate_simulation<-'yes'} indicating no calculation of Jacobian diagnostics}

\item{classvar}{character vector of user specified spatially contiguous discrete
classification variables from sparrow_control.  First element is reach classification variable.}

\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}

\item{numsites}{number of sites selected for calibration}

\item{JacobResults}{list output of Jacobian first-order partial derivatives of the model
residuals \code{estimateNLLSmetrics.R} contained in the estimate.list object.  For more details see
documentation Section 5.2.4.5.}

\item{HesResults}{list output of Hessian second-order standard errors
\code{estimateNLLSmetrics.R} contained in the estimate.list object.  For more details see documentation Section 5.2.4.6}

\item{sparrowEsts}{list object contained in estimate.list \code{if_estimate<-'yes'}.  For more
details see documentation Section 5.2.4.4.}

\item{if_validate}{yes/no indicating whether or not validation is run}

\item{betavalues}{data.frame of model parameters from parameters.csv}
}
\description{
Outputs all model performance and other summary metrics and diagnostics for the
estimated model for calibration and validates sites to the ~/estimate/(run_id)_summary.txt
file.  \cr \cr
Executed By: estimate.R \cr
Executes Routines: unPackList.R \cr
}
