% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWeightedErrors.R
\name{estimateWeightedErrors}
\alias{estimateWeightedErrors}
\title{estimateWeightedErrors}
\usage{
estimateWeightedErrors(file.output.list, xrun_id, pre_run_id, nreaches,
  calsites)
}
\arguments{
\item{xrun_id}{current model run_id used by \code{estimateWeightedErrors.R} executed in
\code{userModifyData.R}}

\item{pre_run_id}{run_id for the prior model residuals used by \code{estimateWeightedErrors.R}
executed in \code{userModifyData.R}}

\item{nreaches}{number of reaches}

\item{calsites}{calibration site indicator (NA or 0 = not selected; 1=selected)}
}
\description{
\cr \cr
Executed By:  \cr
Executes Routines: \itemize{\item named.list.R
\item unPackList.R} \cr
}
