% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMinMaxLatLon.R
\name{findMinMaxLatLon}
\alias{findMinMaxLatLon}
\title{findMinMaxLatLon}
\usage{
findMinMaxLatLon(sitedata, mapping.input.list)
}
\arguments{
\item{sitedata}{Sites selected for calibration using \code{subdata[(subdata$depvar > 0), ]}}
}
\value{
\code{sitegeolimits} named list of minimum and maximum values of the calibration station
latitudes and longitudes
}
\description{
Identifies the minimum and maximum values of the calibration station latitudes
and longitudes, which is output to the Console, and used where the control settings 'lat_limit'
and 'lon_limit' are equal to 'NA' \cr \cr
Executed By: startModelRun.R \cr
Executes Routines: named.list.R \cr
}
