% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBootsOutCSV.R
\name{predictBootsOutCSV}
\alias{predictBootsOutCSV}
\title{predictBootsOutCSV}
\usage{
predictBootsOutCSV(file.output.list, estimate.list, predictBoots.list,
  subdata, add_vars, data_names)
}
\arguments{
\item{subdata}{data.frame input data (subdata)}

\item{add_vars}{additional variables specified by the setting \code{add_vars} to be included in
prediction, yield, and residuals csv and shape files}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}
}
\description{
Outputs the bootstrap predictions to CSV files.  \cr \cr
Executed By: controlFileTasksModel.R \cr
Executes Routines: \itemize{\item getVarList.R
\item unPackList.R} \cr
}
