% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictScenarios.R
\name{predictScenarios}
\alias{predictScenarios}
\title{predictScenarios}
\usage{
predictScenarios(input, allMetrics, output_map_type, Rshiny,
  estimate.input.list, predict.list, scenario.input.list, data_names,
  JacobResults, if_predict, DataMatrix.list, SelParmValues, subdata,
  file.output.list, add_vars, mapping.input.list, batch_mode,
  RSPARROW_errorOption)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{allMetrics}{character string of all load, yield, uncertainty, and data dictionary
variables to map in shiny batch mode}

\item{output_map_type}{character string control setting to identify type of map(s) to output
to PDF file from "stream","catchment", or "both"}

\item{Rshiny}{TRUE/FALSE indicating whether routine is being run from the Shiny app}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{JacobResults}{list output of Jacobian first-order partial derivatives of the model
residuals \code{estimateNLLSmetrics.R} contained in the estimate.list object.  For more details see
documentation Section 5.2.4.5.}

\item{if_predict}{yes/no indicating whether or not prediction is run}

\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{subdata}{data.frame input data (subdata)}

\item{add_vars}{additional variables specified by the setting \code{add_vars} to be included in
prediction, yield, and residuals csv and shape files}

\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}

\item{RSPARROW_errorOption}{}
}
\description{
Executes tasks for the management source-change scenarios, including generating
load predictions for the scenario, saving the predictions to the 'predictScenarios.list', and
executing the mapping function 'predictMaps'. \cr \cr
Executed By: \itemize{\item interactiveBatchRun.R
\item controlFileTasksModel.R
\item goShinyPlot.R} \cr
Executes Routines: \itemize{\item batchMaps.R
\item named.list.R
\item outputSettings.R
\item predictMaps.R
\item predictScenariosOutCSV.R
\item predictScenariosPrep.R
\item unPackList.R
\item deliv_fraction.for
\item mptnoder.for
\item ptnoder.for} \cr
}
