% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictScenariosPrep.R
\name{predictScenariosPrep}
\alias{predictScenariosPrep}
\title{predictScenariosPrep}
\usage{
predictScenariosPrep(input, allMetrics, output_map_type, Rshiny,
  scenario.input.list, data_names, if_predict, data, srcvar, jsrcvar,
  dataNames, JacobResults, subdata, file.output.list)
}
\arguments{
\item{input}{top level interactive user input in Shiny app}

\item{allMetrics}{character string of all load, yield, uncertainty, and data dictionary
variables to map in shiny batch mode}

\item{output_map_type}{character string control setting to identify type of map(s) to output
to PDF file from "stream","catchment", or "both"}

\item{Rshiny}{TRUE/FALSE indicating whether routine is being run from the Shiny app}

\item{data_names}{data.frame of variable metadata from data_Dictionary.csv file}

\item{if_predict}{yes/no indicating whether or not prediction is run}

\item{data}{\code{DataMatrix.list$data} to be used in source reduction scenario setup. For more
details see documentation Section 5.2.3.1}

\item{srcvar}{\code{SelParmValues$srcvar} to be used in source reduction scenario setup. For more
details see documentation Section 5.2.3.2}

\item{jsrcvar}{\code{DataMatrix.list$data.index.list$jsrcvar}  to be used in source reduction
scenario setup. For more details see documentation Section 5.2.3.1}

\item{dataNames}{the names of the FIXED and REQUIRED system variables and explanatory
variables associated with the user-specified model.}

\item{JacobResults}{list output of Jacobian first-order partial derivatives of the model
residuals \code{estimateNLLSmetrics.R} contained in the estimate.list object.  For more details see
documentation Section 5.2.4.5.}

\item{subdata}{data.frame input data (subdata)}
}
\value{
\code{scenarioPrep.list} list of source change scenario setup control settings, data with
change factors applied, and the scenarioFlag indicating the scenario mapping area
}
\description{
Sets up source reduction scenario applying change factors to specified areas \cr \cr
Executed By: predictScenarios.R \cr
Executes Routines: \itemize{\item hydseqTerm.R
\item named.list.R
\item unPackList.R} \cr
}
