% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSummaryOutCSV.R
\name{predictSummaryOutCSV}
\alias{predictSummaryOutCSV}
\title{predictSummaryOutCSV}
\usage{
predictSummaryOutCSV(file.output.list, estimate.input.list, SelParmValues,
  estimate.list, predict.list, subdata, class.input.list)
}
\arguments{
\item{SelParmValues}{selected parameters from parameters.csv using condition
\code{ifelse((parmMax > 0 | (parmType=="DELIVF" & parmMax>=0)) & (parmMin<parmMax) & ((parmType=="SOURCE" &  parmMin>=0) | parmType!="SOURCE")}}

\item{subdata}{data.frame input data (subdata)}
}
\description{
Calculates and outputs to the ~/estimate/(run_id)_summary_predictions.csv file
the percentile summaries of reach predictions of load, yield, concentration, and delivery
fraction for the control settings if_estimate<-"yes" and if_predict<-"yes", and outputs percentile
summaries for the yields  by predominant land-use type specified in the 'class_landuse' control setting. \cr \cr
Executed By: estimate.R \cr
Executes Routines: \itemize{\item calcClassLandusePercent.R
\item unPackList.R} \cr
}
