% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchShiny.R
\name{runBatchShiny}
\alias{runBatchShiny}
\title{runBatchShiny}
\usage{
runBatchShiny(path_PastResults, path_shinyBrowser = NA)
}
\arguments{
\item{path_PastResults}{character string path to previously generated results including run_id subdirectory}

\item{path_shinyBrowser}{character string path to browser executible, if NA defualt browser will be used}
}
\description{
runs shiny app on previously executed RSPARROW model in batch mode from user
specified browser \cr \cr
Executes Routines: shinyBatch.R \cr
}
