% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMaps.R
\name{setupMaps}
\alias{setupMaps}
\title{setupMaps}
\usage{
setupMaps(file.output.list, mapping.input.list, batch_mode,
  RSPARROW_errorOption)
}
\arguments{
\item{batch_mode}{yes/no character string indicating whether RSPARROW is being run in batch
mode}

\item{RSPARROW_errorOption}{}
}
\description{
A preprocessor to read shape files and store objects \cr \cr
Executed By: executeRSPARROW.R \cr
Executes Routines: \itemize{\item batchGeoLines.R
\item batchlineShape.R
\item batchpolyShape.R
\item outputSettings.R
\item unPackList.R} \cr
}
