% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyScenarios.R
\name{shinyScenarios}
\alias{shinyScenarios}
\title{shinyScenarios}
\usage{
shinyScenarios(id, input, choices)
}
\arguments{
\item{id}{Shiny namespace designation}

\item{input}{top level interactive user input in Shiny app}

\item{choices}{data.frame output of function createInteractiveChoices.R}
}
\description{
shiny ui for source change scenario mapping \cr \cr
Executed By: shinyMap2.R \cr
Executes Routines: \itemize{\item dropFunc.R
\item handsOnUI.R} \cr
}
