% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySiteAttr.R
\name{shinySiteAttr}
\alias{shinySiteAttr}
\title{shinySiteAttr}
\usage{
shinySiteAttr(id, input, choices)
}
\arguments{
\item{id}{Shiny namespace designation}

\item{input}{top level interactive user input in Shiny app}

\item{choices}{data.frame output of function createInteractiveChoices.R}
}
\description{
Shiny ui function generates user selections for Site Attribute Mapping Uses
subroutines: dropFunc. \cr \cr
Executed By: shinyMap2.R \cr
Executes Routines: \itemize{\item dropFunc.R
\item handsOnUI.R} \cr
}
